Convert a segmenter to a model object

Convert segmenters into models or standardized segmenters

Usage

# S3 method for class 'tidycpt'
as.model(object, ...)

# S3 method for class 'tidycpt'
as.segmenter(object, ...)

as.segmenter(object, ...)

as.seg_cpt(object, ...)

as.model(object, ...)

# Default S3 method
as.model(object, ...)

is_model(x, ...)

is_segmenter(object, ...)

Arguments

object

A segmenter object, typically returned by segment()

...

currently ignored

x

An object, typically returned by fit_*()

Value

as.segmenter() returns a segmenter object

as.seg_cpt() returns a seg_cpt object

as.model() returns a mod_cpt model object

is_model() a logical vector of length 1

is_segmenter() a logical vector of length 1

Details

as.model.tidycpt() returns the model object of a tidycpt object

as.segmenter.tidycpt() returns the segmenter object of a tidycpt object

as.segmenter() returns the segmenter of a tidycpt object

as.seg_cpt() converts a wild-caught segmenter of any class into a seg_cpt object

as.model() converts a segmenter object into a mod_cpt model object

is_model() checks to see if a model object implements all of the S3 methods necessary to be considered a model.

is_segmenter() checks to see if a segmenter object implements all of the S3 methods necessary to be considered a segmenter.

Examples

cpt <- segment(DataCPSim, method = "ga", maxiter = 5)
as.model(cpt$segmenter)
#> List of 6
#>  $ data         : Time-Series [1:1096] from 1 to 1096: 35.5 29 35.6 33 29.5 ...
#>  $ tau          : int [1:511] 2 3 4 5 7 8 11 14 17 19 ...
#>  $ region_params: tibble [512 × 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ region  : chr [1:512] "[0,2)" "[2,3)" "[3,4)" "[4,5)" ...
#>   ..$ param_mu: num [1:512] 35.5 29 35.6 33 27.5 ...
#>  $ model_params : Named num 273
#>   ..- attr(*, "names")= chr "sigma_hatsq"
#>  $ fitted_values: num [1:1096] 35.5 29 35.6 33 27.5 ...
#>  $ model_name   : chr "meanshift_norm"
#>  - attr(*, "class")= chr "mod_cpt"
if (FALSE) { # \dontrun{
cpt <- segment(DataCPSim, method = "ga-coen", model_fn_args = list(threshold = 80), maxiter = 5)
as.model(cpt$segmenter)$model_params
} # }
x <- fit_nhpp(CET, tau = 330)
is_model(x)
#> [1] TRUE
x <- segment(CET, method = "pelt")
is_segmenter(x$segmenter)
#> [1] TRUE