Retrieve the arguments that a model-fitting function used

Retrieve the arguments that a model-fitting function used

Usage

model_args(object, ...)

# S3 method for default
model_args(object, ...)

# S3 method for ga
model_args(object, ...)

# S3 method for cpt
model_args(object, ...)

Arguments

object

An object, typically returned by segment()

...

currently ignored

Value

model_args(): A named list of arguments, or NULL

Details

Every segmenter uses a model-fitting function, and these functions sometimes take arguments. model_args() recovers the arguments that were passed to the model fitting function when it was called. These are especially important when using a genetic algorithm.

See also

Examples

x <- segment(CET, method = "ga-taimal", maxiter = 3)
#> method: ga-taimal
#> method: pelt
#> method: binseg
#> method: wbs
#> Seeding initial population with probability: 0.0119705340699816
model_args(x$segmenter)
#> $model_fn
#> [1] "nhpp"
#> 
#> $penalty_fn
#> [1] "BMDL"
#>