Segment a time series using a genetic algorithm

Segmenting functions for various genetic algorithms

Usage

segment_ga(
  x,
  model_fn = fit_meanshift_norm,
  penalty_fn = BIC,
  model_fn_args = list(),
  ...
)

segment_ga_shi(x, ...)

segment_ga_coen(x, ...)

segment_ga_random(x, ...)

Arguments

x

A time series

model_fn

A character or name coercible into a fun_cpt() function. See, for example, fit_meanshift_norm().

penalty_fn

A function that evaluates the changepoint set returned by model_fn. We provide AIC(), BIC(), MBIC(), MDL(), and BMDL().

model_fn_args

A list() of parameters passed to model_fn

...

arguments passed to GA::ga()

Details

segment_ga_shi(): Shi's algorithm is the algorithm used in doi:10.1175/JCLI-D-21-0489.1 . Note that in order to achieve the reported results you have to run the algorithm for a really long time. Pass the values maxiter = 50000 and run = 10000 to GA::ga() using the dots.

segment_ga_coen(): Coen's algorithm is the one used in doi:10.1007/978-3-031-47372-2_20 . Note that the speed of the algorithm is highly sensitive to the size of the changepoint sets under consideration, with large changepoint sets being slow. Consider setting the population argument to GA::ga() to improve performance. Coen's algorithm uses the build_gabin_population() function for this purpose by default.

segment_ga_random(): Randomly select candidate changepoint sets. This is implemented as a genetic algorithm with only one generation (i.e., maxiter = 1. Note that this function uses log_gabin_population() by default.

References

Shi, et al. (2022, doi:10.1175/JCLI-D-21-0489.1 )

Taimal, et al. (2023, doi:10.1007/978-3-031-47372-2_20 )

See also

Examples

if (FALSE) {
res <- segment_ga(DataCPSim, maxiter = 10)
summary(res)
str(res)
plot(res)
# Shi's algorithm
x <- segment(CET, method = "ga-shi", maxiter = 50)
str(x)
# GeneticBMDL
y <- segment(CET, method = "ga-coen", maxiter = 20)
changepoints(y)

z <- segment(CET, method = "ga-coen", maxiter = 10, model_fn_args = list(threshold = 2))
changepoints(z)
}
if (FALSE) {
# This will take a really long time!
x <- segment(CET, method = "ga-shi", maxiter = 500, run = 100)
changepoints(x)

y <- segment(CET, method = "ga", model_fn = fit_lmshift, penalty_fn = BIC, 
  popSize = 200, maxiter = 5000, run = 1000, model_fn_args = list(trends = TRUE), 
  population = build_gabin_population(CET))

}

if (FALSE) {
x <- segment(method = "ga-coen", maxiter = 50)
}

x <- segment(CET, method = "random")
#> Seeding initial population with probability: 0.0162752602536624