Convert a segmenter to a model object

Convert segmenters into models or standardized segmenters

Usage

as.segmenter(object, ...)

as.model(object, ...)

# S3 method for default
as.model(object, ...)

is_model(x, ...)

is_segmenter(object, ...)

Arguments

object

A segmenter object, typically returned by segment()

...

currently ignored

x

An object, typically returned by fit_*()

Value

as.segmenter() returns a seg_cpt object

as.model() returns a mod_cpt model object

is_model() a logical vector of length 1

is_segmenter() a logical vector of length 1

Details

as.segmenter() converts a wild-caught segmenter of any class into a seg_cpt object

as.model() converts a segmenter object into a mod_cpt model object

is_model() checks to see if a model object implements all of the S3 methods necessary to be considered a model.

is_segmenter() checks to see if a segmenter object implements all of the S3 methods necessary to be considered a segmenter.

Examples

cpt <- segment(DataCPSim, method = "ga", maxiter = 5)
#> method: ga
as.model(cpt$segmenter)
#> List of 6
#>  $ data         : Time-Series [1:1096] from 1 to 1096: 35.5 29 35.6 33 29.5 ...
#>  $ tau          : int [1:500] 2 3 6 8 10 11 12 13 14 16 ...
#>  $ region_params: tibble [501 × 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ region  : chr [1:501] "[0,2)" "[2,3)" "[3,6)" "[6,8)" ...
#>   ..$ param_mu: num [1:501] 35.5 29 32.7 27.1 37.6 ...
#>  $ model_params : Named num 292
#>   ..- attr(*, "names")= chr "sigma_hatsq"
#>  $ fitted_values: num [1:1096] 35.5 29 32.7 32.7 32.7 ...
#>  $ model_name   : chr "meanshift_norm"
#>  - attr(*, "class")= chr "mod_cpt"
cpt <- segment(DataCPSim, method = "ga-taimal", model_fn_args = list(threshold = 80), maxiter = 5)
#> method: ga-taimal
#> method: pelt
#> method: binseg
#> method: wbs
#> Seeding initial population with probability: 0.0191605839416058
as.model(cpt$segmenter)$model_params
#> threshold 
#>        80 
x <- fit_nhpp(CET, tau = 330)
is_model(x)
#> [1] TRUE
x <- segment(CET, method = "pelt")
#> method: pelt
is_segmenter(x$segmenter)
#> [1] TRUE