Extract changepoints

Retrieve the indices of the changepoints identified by an algorithm or model.

Usage

changepoints(x, ...)

# Default S3 method
changepoints(x, ...)

# S3 method for class 'mod_cpt'
changepoints(x, ...)

# S3 method for class 'seg_basket'
changepoints(x, ...)

# S3 method for class 'seg_cpt'
changepoints(x, ...)

# S3 method for class 'tidycpt'
changepoints(x, use_labels = FALSE, ...)

# S3 method for class 'ga'
changepoints(x, ...)

# S3 method for class 'cpt'
changepoints(x, ...)

# S3 method for class 'wbs'
changepoints(x, ...)

Arguments

x

A tidycpt, segmenter, or mod_cpt object

...

arguments passed to methods

use_labels

return the time labels for the changepoints instead of the indices.

Value

a numeric vector of changepoint indices, or, if use_labels is TRUE, a character of time labels.

Details

tidycpt objects, as well as their segmenter and model components, implement changepoints() methods.

Note that this function is not to be confused with wbs::changepoints(), which returns different information.

For the default method, changepoints() will attempt to return the cpt_true attribute, which is set by test_set().

See also

Examples

cpts <- segment(DataCPSim, method = "ga", maxiter = 5)
changepoints(cpts$segmenter)
#>    x2    x4    x7    x9   x10   x11   x15   x17   x18   x20   x21   x22   x23 
#>     2     4     7     9    10    11    15    17    18    20    21    22    23 
#>   x24   x27   x29   x32   x35   x36   x37   x38   x42   x43   x44   x49   x50 
#>    24    27    29    32    35    36    37    38    42    43    44    49    50 
#>   x54   x55   x60   x61   x62   x68   x69   x70   x71   x78   x80   x81   x83 
#>    54    55    60    61    62    68    69    70    71    78    80    81    83 
#>   x84   x85   x86   x89   x90   x93   x98  x102  x105  x107  x108  x114  x115 
#>    84    85    86    89    90    93    98   102   105   107   108   114   115 
#>  x119  x122  x124  x125  x128  x129  x130  x132  x135  x136  x137  x138  x142 
#>   119   122   124   125   128   129   130   132   135   136   137   138   142 
#>  x143  x145  x152  x158  x159  x162  x163  x167  x170  x175  x176  x177  x180 
#>   143   145   152   158   159   162   163   167   170   175   176   177   180 
#>  x181  x184  x185  x187  x193  x194  x197  x198  x200  x201  x202  x203  x205 
#>   181   184   185   187   193   194   197   198   200   201   202   203   205 
#>  x210  x212  x215  x216  x218  x220  x222  x232  x234  x235  x236  x237  x240 
#>   210   212   215   216   218   220   222   232   234   235   236   237   240 
#>  x242  x244  x246  x247  x249  x250  x255  x258  x262  x263  x264  x271  x273 
#>   242   244   246   247   249   250   255   258   262   263   264   271   273 
#>  x276  x278  x280  x283  x285  x286  x288  x289  x291  x295  x296  x297  x299 
#>   276   278   280   283   285   286   288   289   291   295   296   297   299 
#>  x300  x301  x303  x304  x308  x309  x315  x317  x318  x321  x323  x325  x326 
#>   300   301   303   304   308   309   315   317   318   321   323   325   326 
#>  x329  x330  x331  x334  x336  x337  x339  x342  x343  x344  x346  x347  x348 
#>   329   330   331   334   336   337   339   342   343   344   346   347   348 
#>  x349  x350  x351  x352  x354  x356  x359  x360  x361  x363  x364  x369  x373 
#>   349   350   351   352   354   356   359   360   361   363   364   369   373 
#>  x378  x379  x382  x383  x385  x386  x389  x390  x392  x394  x397  x398  x399 
#>   378   379   382   383   385   386   389   390   392   394   397   398   399 
#>  x401  x402  x403  x405  x408  x409  x410  x413  x414  x415  x416  x420  x423 
#>   401   402   403   405   408   409   410   413   414   415   416   420   423 
#>  x424  x430  x433  x434  x439  x441  x442  x443  x445  x446  x447  x449  x450 
#>   424   430   433   434   439   441   442   443   445   446   447   449   450 
#>  x451  x453  x454  x455  x456  x457  x461  x463  x466  x467  x470  x474  x478 
#>   451   453   454   455   456   457   461   463   466   467   470   474   478 
#>  x481  x482  x483  x489  x490  x491  x503  x504  x506  x509  x510  x511  x512 
#>   481   482   483   489   490   491   503   504   506   509   510   511   512 
#>  x515  x517  x518  x520  x524  x527  x530  x532  x536  x537  x546  x547  x550 
#>   515   517   518   520   524   527   530   532   536   537   546   547   550 
#>  x552  x554  x555  x558  x560  x564  x565  x570  x572  x575  x576  x577  x579 
#>   552   554   555   558   560   564   565   570   572   575   576   577   579 
#>  x580  x585  x586  x587  x588  x590  x593  x596  x597  x599  x600  x601  x602 
#>   580   585   586   587   588   590   593   596   597   599   600   601   602 
#>  x607  x608  x609  x613  x617  x618  x619  x620  x621  x624  x625  x628  x629 
#>   607   608   609   613   617   618   619   620   621   624   625   628   629 
#>  x631  x633  x638  x639  x641  x643  x645  x647  x649  x652  x653  x654  x656 
#>   631   633   638   639   641   643   645   647   649   652   653   654   656 
#>  x658  x659  x660  x663  x664  x665  x666  x670  x673  x674  x675  x678  x679 
#>   658   659   660   663   664   665   666   670   673   674   675   678   679 
#>  x680  x682  x684  x686  x690  x691  x692  x694  x696  x702  x704  x715  x718 
#>   680   682   684   686   690   691   692   694   696   702   704   715   718 
#>  x720  x725  x727  x729  x731  x732  x735  x736  x737  x739  x742  x747  x748 
#>   720   725   727   729   731   732   735   736   737   739   742   747   748 
#>  x749  x750  x751  x757  x759  x760  x761  x764  x771  x773  x774  x777  x778 
#>   749   750   751   757   759   760   761   764   771   773   774   777   778 
#>  x779  x788  x793  x794  x795  x798  x799  x800  x801  x802  x804  x805  x806 
#>   779   788   793   794   795   798   799   800   801   802   804   805   806 
#>  x807  x808  x809  x811  x812  x814  x815  x816  x818  x820  x821  x822  x827 
#>   807   808   809   811   812   814   815   816   818   820   821   822   827 
#>  x828  x832  x835  x836  x841  x843  x844  x848  x850  x851  x852  x853  x855 
#>   828   832   835   836   841   843   844   848   850   851   852   853   855 
#>  x856  x857  x858  x859  x861  x863  x866  x868  x869  x871  x872  x873  x874 
#>   856   857   858   859   861   863   866   868   869   871   872   873   874 
#>  x876  x879  x880  x882  x886  x889  x893  x899  x900  x902  x903  x910  x912 
#>   876   879   880   882   886   889   893   899   900   902   903   910   912 
#>  x914  x918  x919  x921  x926  x927  x928  x929  x930  x931  x938  x939  x941 
#>   914   918   919   921   926   927   928   929   930   931   938   939   941 
#>  x943  x944  x946  x948  x951  x952  x953  x954  x956  x957  x958  x959  x966 
#>   943   944   946   948   951   952   953   954   956   957   958   959   966 
#>  x967  x968  x969  x970  x972  x974  x975  x977  x981  x982  x983  x986  x988 
#>   967   968   969   970   972   974   975   977   981   982   983   986   988 
#>  x989  x991  x992  x995  x996  x998 x1001 x1002 x1003 x1004 x1007 x1008 x1010 
#>   989   991   992   995   996   998  1001  1002  1003  1004  1007  1008  1010 
#> x1014 x1015 x1023 x1024 x1025 x1027 x1028 x1030 x1033 x1037 x1041 x1042 x1047 
#>  1014  1015  1023  1024  1025  1027  1028  1030  1033  1037  1041  1042  1047 
#> x1048 x1050 x1052 x1053 x1055 x1060 x1064 x1065 x1066 x1069 x1072 x1073 x1075 
#>  1048  1050  1052  1053  1055  1060  1064  1065  1066  1069  1072  1073  1075 
#> x1076 x1079 x1081 x1082 x1083 x1085 x1086 x1088 x1090 x1092 x1094 x1095 
#>  1076  1079  1081  1082  1083  1085  1086  1088  1090  1092  1094  1095 

cpts <- segment(DataCPSim, method = "wbs")
changepoints(cpts$segmenter)
#>  [1]  547  822  939  972  981  982  997  999 1031 1033 1040 1041 1063 1066