Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    7    1  613  629  657  791  888 1014 1086  1096     0     0     0
#>  [2,]   13    1  415  589  680  774  834  900  911   953   954   961  1034
#>  [3,]    5    1  681  813  928  972 1038 1096    0     0     0     0     0
#>  [4,]    7    1  746  951  988 1010 1020 1024 1068  1096     0     0     0
#>  [5,]    8    1  415  780  833  837  895  932  968  1010  1096     0     0
#>  [6,]    7    1  605  729  780 1006 1024 1026 1086  1096     0     0     0
#>  [7,]   10    1  575  604  646  839  870  905  953   993  1068  1086  1096
#>  [8,]    9    1  634  672  863  895  915  950 1028  1051  1086  1096     0
#>  [9,]    9    1  488  590  899  931 1008 1016 1073  1075  1086  1096     0
#> [10,]    9    1  634  678  894  896  943  998 1044  1055  1068  1096     0
#> [11,]    9    1  681  763  780  953  987  997 1037  1040  1085  1096     0
#> [12,]    9    1  646  834  871  905  975  977  991  1071  1086  1096     0
#> [13,]    9    1  583  678  697  864  883  894 1004  1006  1077  1096     0
#> [14,]    8    1  655  774  937  944  951  968 1005  1086  1096     0     0
#> [15,]   13    1  657  671  852  885  903  920  939   955   980  1004  1015
#> [16,]    8    1  619  672  706  842  935  977 1039  1056  1096     0     0
#> [17,]   10    1  604  791  857  890  894  943  966   995  1017  1062  1096
#> [18,]   10    1  659  688  832  852  896  904  931   933  1023  1034  1096
#> [19,]    7    1  823  824  868  979  990 1077 1084  1096     0     0     0
#> [20,]    7    1  634  705  839  882  883 1020 1080  1096     0     0     0
#> [21,]   13    1  657  716  719  839  931  941 1003  1018  1019  1030  1058
#> [22,]   11    1  553  645  716  777  780  841  971   987  1018  1028  1042
#> [23,]    6    1  885 1019 1026 1057 1073 1083 1096     0     0     0     0
#> [24,]    9    1  553  609  654  746  774  951  985  1050  1086  1096     0
#> [25,]    9    1  706  755  894  954 1042 1049 1053  1059  1081  1096     0
#> [26,]    9    1  642  842  890  894  951  972 1025  1043  1054  1096     0
#> [27,]   12    1  415  589  609  633  656  853  885   901   910   954  1029
#> [28,]    9    1  415  613  794  853  864  932  953  1007  1086  1096     0
#> [29,]   10    1  634  659  694  733  755  763  852   885   935  1011  1096
#> [30,]   11    1  415  672  687  711  725  837  994  1014  1034  1036  1056
#> [31,]   10    1  553  598  785  843  852  863  889   914   934  1013  1096
#> [32,]    6    1  633  763  864  883 1030 1086 1096     0     0     0     0
#> [33,]   11    1  557  654  824  987  997 1019 1020  1027  1043  1055  1083
#> [34,]    6    1  415  604  798  858 1005 1039 1096     0     0     0     0
#> [35,]    7    1  415  852  929  933  954 1042 1070  1096     0     0     0
#> [36,]    6    1  657  738  792  885  918 1048 1096     0     0     0     0
#> [37,]   11    1  415  556  748  791  798  857  924   948   960   983  1043
#> [38,]    9    1  671  678  841  883  929  950  962   963  1086  1096     0
#> [39,]   10    1  415  763  790  791  854  885  892  1045  1046  1063  1096
#> [40,]    7    1  609  672  792  883  953 1029 1059  1096     0     0     0
#> [41,]    9    1  575  672  822  907  919  970 1017  1053  1062  1096     0
#> [42,]   11    1  604  680  746  813  903  981  985  1003  1041  1048  1069
#> [43,]   12    1  678  706  748  774  832  844  864   866   892   894   926
#> [44,]    9    1  488  763  864  900  919 1048 1064  1072  1086  1096     0
#> [45,]    6    1  861  924  951  966 1018 1045 1096     0     0     0     0
#> [46,]    5    1  646  697  912  919 1020 1096    0     0     0     0     0
#> [47,]    9    1  415  562  575  697  780  870  886   890   945  1096     0
#> [48,]    7    1  672  749  755  868 1032 1038 1078  1096     0     0     0
#> [49,]    8    1  488  633  657  832  854  863  896   912  1096     0     0
#> [50,]    4    1  680  705  892 1086 1096    0    0     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]     0     0     0     0     0     0     0
#>  [2,]  1046  1063  1096     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]  1072  1078  1096     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]  1059  1066  1096     0     0     0     0
#> [22,]  1096     0     0     0     0     0     0
#> [23,]     0     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]  1076  1096     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]  1096     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]  1096     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]  1096     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]  1096     0     0     0     0     0     0
#> [43,]  1005  1096     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  308  553  733  785  831  868  887  901  928  964  981  997 1005
#> [16] 1020 1049 1062 1085 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1  134  314  452  476  709  841  848  854  857  877  884  901  957
#> [16]  963  979 1038 1055 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  415  439  525  633  683  724  748  749  770  783  810  928  934
#> [16]  936  942 1001 1070 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1    5  216  676  766  768  838  868  871  876  902  927  962  971
#> [16]  994 1030 1034 1058 1096
sim_1_cp_BMDL(exceedances(bogota_pm))
#>  [1]   17    1   31   51  164  215  228  248  274  318  341  376  632  663  668
#> [16]  710  820 1038 1050 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  605  680  705  717  728  729  774   826   864   865   870
#>  [2,]   17    1  725  765  843  860  889  912  924   950   953  1031  1052
#>  [3,]   17    1  226  619  794  828  830  861  862   882   888   919   941
#>  [4,]   17    1  354  671  749  755  825  871  889   933   937   961   971
#>  [5,]   17    1  656  713  780  790  792  798  857   877   881   882   943
#>  [6,]   17    1  308  729  823  853  858  861  885   926   939   947   972
#>  [7,]   17    1  604  677  694  716  729  748  753   771   842   891   908
#>  [8,]   16    1  556  677  713  716  843  866  885   887   893   934   972
#>  [9,]   16    1  589  619  649  659  698  749  860   861   893   929   930
#> [10,]   17    1  562  571  604  677  798  824  825   831   854   868   875
#> [11,]   17    1  589  621  647  698  728  738  822   824   830   856   915
#> [12,]   17    1  354  557  561  647  656  678  704   741   882   915   947
#> [13,]   17    1  271  645  656  688  838  844  845   853   861   886   905
#> [14,]   17    1  221  571  633  677  717  719  727   823   842   860   891
#> [15,]   17    1  605  615  619  655  832  836  839   866   883   933   943
#> [16,]   17    1  488  598  618  649  687  698  717   813   817   848   861
#> [17,]   17    1  642  728  785  798  851  883  907   908   931   957   983
#> [18,]   14    1  548  677  698  711  784  849  853   857   915   940   976
#> [19,]   17    1  589  634  722  774  830  858  861   875   926   934   946
#> [20,]   16    1  548  566  613  722  753  883  901   931   945   966   967
#> [21,]   17    1  575  614  623  704  729  792  837   880   889   899   946
#> [22,]   17    1  561  562  753  844  849  919  928   939   940   975   976
#> [23,]   17    1   51  671  706  708  713  716  741   843   884   900   903
#> [24,]   17    1  226  619  628  688  790  798  833   865   923   924  1008
#> [25,]   17    1  590  657  672  727  822  830  831   910   950   951   970
#> [26,]   17    1  596  629  678  823  848  868  887   893   898   901  1019
#> [27,]   17    1  571  581  728  803  817  853  911   924   926   929   939
#> [28,]   17    1  655  803  864  877  882  887  928   931   940   949   969
#> [29,]   17    1  633  803  840  860  882  883  944   963   976   990  1005
#> [30,]   17    1  557  566  655  688  749  774  885   916   933   968   979
#> [31,]   17    1  822  830  833  845  854  866  875   881   895   905   912
#> [32,]   17    1   60  233  613  668  677  705  722   733   901   962   972
#> [33,]   17    1  557  621  628  738  774  837  840   852   861   875   889
#> [34,]   17    1  615  656  727  741  845  852  853   880   912   918   937
#> [35,]   17    1  415  694  719  725  830  851  867   877   896   932   986
#> [36,]   17    1  557  619  719  729  733  746  777   809   848   945   981
#> [37,]   17    1  571  596  708  813  830  840  882   888   895   903   969
#> [38,]   17    1  221  308  634  647  677  716  727   755   771   790   833
#> [39,]   17    1  562  668  681  722  833  849  863   877   890   923   933
#> [40,]   17    1  548  656  668  837  851  867  868   875   888   896   968
#> [41,]   17    1  233  753  761  765  846  865  866   935   939   945   984
#> [42,]   17    1  221  633  645  657  681  719  780   817   884   900   934
#> [43,]   17    1  571  575  645  741  885  903  911   933   959   964   967
#> [44,]   17    1  581  596  642  680  729  748  809   836   854   861   894
#> [45,]   17    1  589  598  623  642  663  729  761   790   930  1010  1012
#> [46,]   17    1  596  681  729  828  842  852  854   859   866   924   945
#> [47,]   17    1  596  677  822  861  863  877  905   919   963   964   967
#> [48,]   17    1  226  271  415  618  755  785  831   882   883   924   948
#> [49,]   17    1   20  654  681  687  711  748  763   771   823   892   928
#> [50,]   17    1  575  663  681  837  841  854  859   867   891   932   964
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   936   940   954   983  1024  1081  1096
#>  [2,]  1053  1057  1064  1070  1077  1089  1096
#>  [3,]   979  1012  1033  1040  1054  1090  1096
#>  [4,]   995  1018  1046  1072  1078  1086  1096
#>  [5,]   946   957   966   968  1014  1045  1096
#>  [6,]   995  1004  1057  1061  1077  1094  1096
#>  [7,]   939   940  1019  1032  1036  1048  1096
#>  [8,]   985  1011  1021  1044  1092  1096     0
#>  [9,]   980  1022  1028  1030  1071  1096     0
#> [10,]   885  1004  1045  1071  1074  1078  1096
#> [11,]   930   947   960  1024  1036  1069  1096
#> [12,]   951  1031  1045  1046  1063  1064  1096
#> [13,]   968   971   997  1017  1022  1081  1096
#> [14,]   937   961   985  1019  1080  1092  1096
#> [15,]   951   969  1041  1073  1074  1091  1096
#> [16,]   878   904   947   954   966  1019  1096
#> [17,]  1014  1019  1025  1034  1036  1074  1096
#> [18,]   986  1011  1022  1096     0     0     0
#> [19,]   974   983   993  1001  1004  1093  1096
#> [20,]   979   981  1061  1086  1094  1096     0
#> [21,]   991  1005  1035  1046  1052  1088  1096
#> [22,]   978   992  1023  1040  1066  1083  1096
#> [23,]   939   966   967   980   991  1010  1096
#> [24,]  1022  1027  1029  1036  1037  1079  1096
#> [25,]   984   989  1022  1030  1052  1077  1096
#> [26,]  1025  1032  1067  1082  1087  1088  1096
#> [27,]   978  1000  1004  1022  1024  1077  1096
#> [28,]   990  1005  1017  1036  1076  1082  1096
#> [29,]  1009  1027  1033  1048  1080  1089  1096
#> [30,]   983  1001  1026  1062  1079  1087  1096
#> [31,]   935   959   990  1027  1059  1065  1096
#> [32,]   983  1003  1013  1016  1029  1078  1096
#> [33,]   918   941   966   967   973  1045  1096
#> [34,]   953   986  1000  1014  1063  1069  1096
#> [35,]   987   992  1017  1066  1085  1092  1096
#> [36,]   985  1006  1053  1068  1071  1084  1096
#> [37,]   984  1010  1019  1041  1070  1082  1096
#> [38,]   871   905   931   949  1011  1088  1096
#> [39,]   936   988  1013  1015  1029  1033  1096
#> [40,]  1012  1022  1031  1079  1093  1094  1096
#> [41,]   995  1020  1023  1038  1060  1061  1096
#> [42,]   976   988   995  1028  1030  1075  1096
#> [43,]  1002  1023  1026  1027  1028  1070  1096
#> [44,]   932   954   964   998  1008  1048  1096
#> [45,]  1022  1026  1056  1072  1089  1094  1096
#> [46,]   974   982   998  1042  1061  1074  1096
#> [47,]   989   990  1008  1050  1054  1056  1096
#> [48,]   981   984  1000  1013  1043  1079  1096
#> [49,]   940   941   993  1042  1078  1090  1096
#> [50,]   988   997  1001  1037  1069  1073  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986