Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   11    1  583  678  681  790  831  891  923   941  1052  1078  1086
#>  [2,]    8    1  557  642  657  741  849  896  969  1040  1096     0     0
#>  [3,]    8    1  415  604  834  966  969 1011 1032  1068  1096     0     0
#>  [4,]    9    1  899  943  944  976 1011 1033 1034  1041  1086  1096     0
#>  [5,]    6    1  488  605  856  865  976 1042 1096     0     0     0     0
#>  [6,]    8    1  719  768  845  892  949  984 1061  1085  1096     0     0
#>  [7,]   11    1  605  697  698  837  839  856  859   882   892   905  1072
#>  [8,]    9    1  618  623  748  833  854  904  963  1001  1028  1096     0
#>  [9,]    6    1  655  680  868  976  987  991 1096     0     0     0     0
#> [10,]    9    1  862  885  978 1003 1022 1035 1041  1067  1070  1096     0
#> [11,]   13    1  596  621  655  656  748  771  785   876   904   911   992
#> [12,]    8    1  566  628  647  698  749  891 1042  1055  1096     0     0
#> [13,]    4    1  870  888  953 1008 1096    0    0     0     0     0     0
#> [14,]    8    1  581  704  826  910  972 1003 1014  1082  1096     0     0
#> [15,]    8    1  722  809  850  961 1047 1072 1078  1086  1096     0     0
#> [16,]    8    1  415  746  875  944 1012 1046 1048  1083  1096     0     0
#> [17,]    9    1  706  708  738  850  887  973 1008  1046  1086  1096     0
#> [18,]    8    1  604  615  763  857  879 1009 1057  1085  1096     0     0
#> [19,]    7    1  415  654  705  861  945  969  988  1096     0     0     0
#> [20,]   11    1  488  619  647  656  810  823  833   868   988  1009  1086
#> [21,]    6    1  557  589  824  846 1026 1073 1096     0     0     0     0
#> [22,]    9    1  688  832  858  940  997 1048 1071  1084  1086  1096     0
#> [23,]   11    1  615  708  882  884  963  996 1000  1047  1063  1072  1081
#> [24,]   10    1  556  589  825  833  872  943  962  1006  1073  1074  1096
#> [25,]    8    1  749  826  841  864  883  940 1048  1082  1096     0     0
#> [26,]    9    1  763  931  961  972  975 1010 1044  1051  1057  1096     0
#> [27,]    9    1  548  711  833  883  904  963  992   997  1031  1096     0
#> [28,]    9    1  749  858  859  895  916  979 1018  1075  1083  1096     0
#> [29,]    8    1  583  655  725  785  899  940  941   948  1096     0     0
#> [30,]    6    1  830  851  884  885  961 1086 1096     0     0     0     0
#> [31,]    9    1  605  697  883  945 1009 1063 1067  1072  1086  1096     0
#> [32,]    8    1  553  633  824  831  941  962  987   991  1096     0     0
#> [33,]   11    1  725  780  828  861  905  944  979   992  1012  1060  1082
#> [34,]    5    1  646  655  924  975  979 1096    0     0     0     0     0
#> [35,]    9    1  571  609  780  826  854  944  997  1009  1077  1096     0
#> [36,]   10    1  725  854  891 1000 1015 1026 1027  1029  1067  1086  1096
#> [37,]   11    1  609  738  908  916  936  945 1012  1016  1028  1030  1031
#> [38,]   12    1  655  708  837  872  881  884  892   900   973  1037  1038
#> [39,]    9    1  613  649  680  817  872  877  907   981  1047  1096     0
#> [40,]    7    1  753  755  844  854  929  978  987  1096     0     0     0
#> [41,]    6    1  605  613  645  896  898  939 1096     0     0     0     0
#> [42,]    7    1  415  657  940  988  992 1003 1057  1096     0     0     0
#> [43,]   10    1  589  598  694  834  860  993 1022  1026  1047  1071  1096
#> [44,]   11    1  687  697  711  771  832  841  920   994  1046  1049  1073
#> [45,]    8    1  628  647  655  830  889  900  919  1005  1096     0     0
#> [46,]    8    1  590  619  705  969 1009 1014 1028  1068  1096     0     0
#> [47,]    3    1  825  830 1083 1096    0    0    0     0     0     0     0
#> [48,]    7    1  548  634  649  824  976 1044 1065  1096     0     0     0
#> [49,]    9    1  596  698  753  768  877  908 1005  1063  1074  1096     0
#> [50,]    6    1  761  861  866 1022 1038 1086 1096     0     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  1096     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]  1096     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]  1009  1086  1096     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]  1096     0     0     0     0     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]  1096     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]  1096     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]  1096     0     0     0     0     0     0
#> [38,]  1086  1096     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0
#> [44,]  1096     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  657  713  791  798  803  817  828  859  876  888  991  994 1032
#> [16] 1034 1041 1045 1049 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   59  296  445  493  515  561  589  689  845  911  951  982  993
#> [16] 1021 1026 1031 1052 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  536  537  538  579  695  750  812  816  834  840  890  909  976
#> [16]  994 1029 1035 1088 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1   82  184  419  551  756  839  859  862  893  896  917  954  974
#> [16] 1000 1022 1037 1045 1096
sim_1_cp_BMDL(exceedances(bogota_pm))
#>  [1]   17    1   24  120  173  255  276  348  382  423  425  632  635  664  785
#> [16]  797  984 1030 1059 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  271  615  633  748  761  777  823   832   849   886   905
#>  [2,]   17    1  488  590  655  716  826  831  872   884   889   892   900
#>  [3,]   17    1  698  746  798  830  856  893  895   908   937   963   979
#>  [4,]   17    1   20  589  590  713  763  841  849   911   918   934   957
#>  [5,]   17    1   60  605  609  614  647  777  817   836   840   903   971
#>  [6,]   17    1  589  604  656  716  727  746  858   898   929   935   953
#>  [7,]   17    1  556  598  629  633  761  785  852   862   891   904   908
#>  [8,]   17    1  654  694  841  843  857  870  890   893   898   900   930
#>  [9,]   17    1  553  656  728  748  822  831  842   885   890   893   929
#> [10,]   17    1  621  738  825  866  904  921  928   932   957   961   979
#> [11,]   17    1  221  548  604  656  680  687  803   830   831   838   872
#> [12,]   17    1  221  621  647  753  768  826  864   949   959   966   974
#> [13,]   17    1  566  609  615  618  646  711  774   794   871   890   897
#> [14,]   17    1   51  271  590  663  817  886  930   953   971   982   992
#> [15,]   17    1  308  354  621  634  705  727  803   854   858   875   891
#> [16,]   17    1  581  654  849  853  859  872  893   918   926   928   940
#> [17,]   17    1   60  604  623  647  719  722  755   780   792   836   863
#> [18,]   17    1  642  677  798  823  848  859  868   876   882   890   894
#> [19,]   17    1  583  681  687  771  832  910  929   933   936   948   978
#> [20,]   17    1  488  598  647  680  749  846  862   868   894   915   949
#> [21,]   17    1  571  613  705  741  803  822  839   849   866   868   933
#> [22,]   17    1   20  548  575  681  688  708  765   784   843   910   919
#> [23,]   15    1  604  609  681  785  803  842  843   870   939   951   988
#> [24,]   17    1  226  561  763  798  809  810  823   865   868   879   900
#> [25,]   17    1  415  628  668  671  708  719  785   825   854   868   934
#> [26,]   17    1  583  655  688  711  840  851  911   988  1000  1012  1013
#> [27,]   17    1  678  688  697  844  871  938  943   950   984   988  1002
#> [28,]   17    1  615  645  694  704  729  741  753   765   817   863   889
#> [29,]   17    1  561  628  657  765  849  882  887   891   912   934   944
#> [30,]   17    1   60  233  354  628  746  882  891   892   911   923   962
#> [31,]   17    1   20  609  859  872  875  953  959   960   984   999  1000
#> [32,]   17    1   51  226  668  677  825  838  870   907   933   936   939
#> [33,]   17    1  548  561  605  688  791  825  836   885   895   920   972
#> [34,]   17    1  604  666  668  704  713  749  785   830   849   889  1009
#> [35,]   17    1  548  668  716  748  810  850  867   935   991   994  1000
#> [36,]   17    1   60  271  548  589  746  780  851   870   883   886   918
#> [37,]   17    1   51  575  590  604  621  708  725   822   834   842   866
#> [38,]   17    1  633  645  719  771  780  798  846   882   912   915   963
#> [39,]   17    1  415  688  706  840  853  877  880   907   937   963   992
#> [40,]   17    1  589  663  668  681  694  741  777   846   870   878   894
#> [41,]   17    1  556  645  646  647  680  681  860   912   940   957   960
#> [42,]   17    1  226  233  596  784  809  813  889   899   904   926   929
#> [43,]   17    1  623  633  656  719  828  877  940   945   949   961   977
#> [44,]   17    1  609  618  681  763  771  885  919   945   955   961   980
#> [45,]   17    1  687  722  794  830  832  837  861   880   891   899   908
#> [46,]   17    1  221  613  633  755  777  826  832   840   897   905   924
#> [47,]   17    1  605  629  659  832  843  866  889   928   938   951   988
#> [48,]   17    1  226  654  671  678  719  803  822   854   858   878   895
#> [49,]   17    1   51  628  711  717  830  844  862   876   896   910   954
#> [50,]   17    1  628  642  645  655  828  862  863   911   943   998  1006
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   915   926   995  1005  1041  1079  1096
#>  [2,]   903   936   957   972  1063  1083  1096
#>  [3,]   987   988  1005  1028  1054  1076  1096
#>  [4,]   990  1010  1052  1067  1071  1093  1096
#>  [5,]   977   980  1009  1031  1047  1064  1096
#>  [6,]   982  1006  1059  1064  1066  1088  1096
#>  [7,]   936   950   970   979  1015  1033  1096
#>  [8,]   944   948  1000  1019  1024  1044  1096
#>  [9,]  1011  1012  1036  1037  1045  1058  1096
#> [10,]  1007  1021  1046  1062  1075  1087  1096
#> [11,]   877   883   885   960  1048  1082  1096
#> [12,]   979   988  1069  1087  1090  1092  1096
#> [13,]   941   985  1019  1062  1073  1077  1096
#> [14,]  1005  1048  1085  1088  1089  1090  1096
#> [15,]   947   986   998  1024  1036  1079  1096
#> [16,]   978  1026  1035  1058  1060  1092  1096
#> [17,]   911   945  1007  1072  1080  1087  1096
#> [18,]   923   935   963  1008  1078  1088  1096
#> [19,]   987   988  1025  1047  1068  1084  1096
#> [20,]  1002  1024  1044  1063  1067  1093  1096
#> [21,]   967   988  1009  1011  1013  1089  1096
#> [22,]   928   930   948   975  1016  1023  1096
#> [23,]   992  1003  1010  1059  1096     0     0
#> [24,]   919   949  1007  1008  1052  1073  1096
#> [25,]   948  1006  1026  1033  1036  1076  1096
#> [26,]  1018  1030  1043  1044  1076  1077  1096
#> [27,]  1015  1021  1033  1044  1065  1091  1096
#> [28,]   932   954   962  1054  1068  1073  1096
#> [29,]   949   972  1002  1030  1041  1078  1096
#> [30,]   964  1017  1018  1020  1076  1080  1096
#> [31,]  1016  1029  1058  1060  1083  1092  1096
#> [32,]   970   999  1015  1024  1045  1078  1096
#> [33,]  1035  1043  1044  1064  1083  1093  1096
#> [34,]  1021  1028  1045  1048  1062  1070  1096
#> [35,]  1009  1052  1060  1064  1070  1086  1096
#> [36,]   997  1023  1029  1037  1081  1087  1096
#> [37,]   897   950   971  1016  1066  1084  1096
#> [38,]   983   992  1008  1013  1043  1093  1096
#> [39,]  1005  1064  1068  1072  1075  1092  1096
#> [40,]   901   989  1019  1072  1092  1095  1096
#> [41,]   982   988  1024  1047  1071  1082  1096
#> [42,]   937   960   987  1009  1054  1072  1096
#> [43,]   982  1023  1026  1041  1056  1095  1096
#> [44,]   987  1047  1064  1071  1084  1091  1096
#> [45,]   950   987  1033  1039  1060  1077  1096
#> [46,]   947  1044  1058  1072  1078  1083  1096
#> [47,]  1035  1060  1065  1072  1075  1078  1096
#> [48,]   915   938   949  1048  1050  1062  1096
#> [49,]   960   993  1009  1013  1075  1092  1096
#> [50,]  1018  1046  1061  1062  1086  1094  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986