Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   11    1  415  705  741  843  867  901  912   923   961  1017  1049
#>  [2,]    9    1  645  865  953  984  994 1013 1029  1058  1086  1096     0
#>  [3,]    7    1  488  859  862  878  945  970 1086  1096     0     0     0
#>  [4,]    8    1  415  596  923  967  992  996 1007  1077  1096     0     0
#>  [5,]    9    1  589  672  687  896  911  936  974  1033  1036  1096     0
#>  [6,]   11    1  655  656  741  886  995 1009 1042  1052  1055  1058  1079
#>  [7,]    8    1  727  729  755  761  824  949 1025  1049  1096     0     0
#>  [8,]   11    1  596  654  655  705  753  846  851   878   963  1001  1014
#>  [9,]   11    1  571  583  654  672  748  893  910   938   971  1058  1073
#> [10,]    7    1  681  842  857  910 1044 1048 1065  1096     0     0     0
#> [11,]    6    1  663  884  932  977 1049 1086 1096     0     0     0     0
#> [12,]    6    1  629  663  698  765 1064 1084 1096     0     0     0     0
#> [13,]   10    1  557  562  688  704  871  916  970   979   987   989  1096
#> [14,]   10    1  663  672  728  839  868  940  950   980   981  1064  1096
#> [15,]    9    1  571  738  887  919  968 1001 1037  1051  1086  1096     0
#> [16,]    7    1  415  646  910 1044 1051 1064 1080  1096     0     0     0
#> [17,]    5    1  655  681  843 1074 1086 1096    0     0     0     0     0
#> [18,]   10    1  415  605  619  655  870  995 1007  1036  1051  1086  1096
#> [19,]   10    1  553  719  817  825  888  896  990  1014  1040  1058  1096
#> [20,]    9    1  722  790  851  860  900  979 1024  1083  1085  1096     0
#> [21,]   12    1  556  598  654  659  688  753  876   893   939   997  1042
#> [22,]    7    1  415  863  871  910  961  983 1086  1096     0     0     0
#> [23,]    9    1  621  704  777  843  870  985 1028  1050  1076  1096     0
#> [24,]    9    1  557  634  694  748  885  887  996  1043  1067  1096     0
#> [25,]   11    1  680  716  719  733  923  924  931   972  1005  1038  1049
#> [26,]   12    1  553  619  655  666  681  859 1002  1014  1035  1038  1082
#> [27,]    8    1  557  590  654  688  716  830  897   997  1096     0     0
#> [28,]   11    1  488  727  824  875  897  933  948   954  1023  1045  1058
#> [29,]    9    1  590  633  649  705  946 1030 1043  1077  1086  1096     0
#> [30,]    8    1  645  877  918  934  994 1014 1043  1044  1096     0     0
#> [31,]   10    1  548  566  629  659  672  706  836  1018  1031  1047  1096
#> [32,]    5    1  687  871 1006 1044 1051 1096    0     0     0     0     0
#> [33,]    6    1  415  919  940  961 1069 1086 1096     0     0     0     0
#> [34,]   13    1  553  562  621  666  688  763  839   875   931  1007  1039
#> [35,]    8    1  771  871  954  992 1034 1045 1058  1076  1096     0     0
#> [36,]    6    1  415  598  604  704  912  961 1096     0     0     0     0
#> [37,]    7    1  415  753  872  931  948 1017 1023  1096     0     0     0
#> [38,]   11    1  604  656  672  790  791  893  919   946  1017  1023  1086
#> [39,]    9    1  609  629  798  875  878  887  950   981  1058  1096     0
#> [40,]    7    1  619  838  840  951  990 1006 1052  1096     0     0     0
#> [41,]    7    1  583  950 1025 1033 1045 1049 1051  1096     0     0     0
#> [42,]   10    1  562  596  615  633  698  896  991   998  1006  1028  1096
#> [43,]   13    1  822  834  838  879  881  887  918   971   992  1027  1051
#> [44,]    9    1  561  566  657  677  794  893  996  1041  1068  1096     0
#> [45,]   11    1  590  621  708  867  898  974 1004  1015  1030  1051  1078
#> [46,]    7    1  488  598  678  791  858  888 1086  1096     0     0     0
#> [47,]    9    1  656  698  738  790  850  881  896   918  1014  1096     0
#> [48,]    4    1  934  944  991 1005 1096    0    0     0     0     0     0
#> [49,]    5    1  645  694  867  969 1068 1096    0     0     0     0     0
#> [50,]    8    1  746  763  837  910 1012 1039 1048  1050  1096     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  1096     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]  1096     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]  1096     0     0     0     0     0     0
#>  [9,]  1096     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]  1066  1096     0     0     0     0     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]     0     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]  1096     0     0     0     0     0     0
#> [26,]  1086  1096     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]  1096     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]  1051  1086  1096     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]  1096     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]  1074  1086  1096     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]  1096     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  649  666  678  716  755  826  883  886  937  947  974  977 1019
#> [16] 1061 1067 1086 1093 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1  126  147  354  446  485  570  706  728  788  832  858  864  894
#> [16]  914  943 1072 1087 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  379  486  637  729  745  759  774  785  799  810  861  862  903
#> [16]  942  947  989 1082 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1  610  650  697  766  845  871  874  876  899  905  955  967  992
#> [16] 1014 1031 1048 1067 1096
sim_1_cp_BMDL(exceedances(bogota_pm))
#>  [1]   17    1   77  102  166  448  452  459  599  604  711  712  807  816  901
#> [16]  983 1005 1033 1085 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  571  590  645  659  688  704  725   774   777   840   879
#>  [2,]   17    1  604  768  825  852  867  870  915   949  1014  1044  1046
#>  [3,]   17    1  557  645  823  825  845  848  852   900   919   943   946
#>  [4,]   17    1  226  415  488  678  706  708  733   768   777   830   932
#>  [5,]   17    1  777  792  798  824  868  898  931   935   938   941   970
#>  [6,]   16    1  415  590  671  777  810  838  868   890   898   905   951
#>  [7,]   17    1  562  598  657  678  716  765  826   850   899   903   908
#>  [8,]   17    1  609  618  629  672  697  698  711   717   774   833   837
#>  [9,]   17    1  221  308  556  566  663  677  711   755   798   830   854
#> [10,]   15    1  562  687  741  834  849  850  860   866   875   911   919
#> [11,]   17    1  233  561  609  719  753  794  848   883   924   953   980
#> [12,]   17    1  548  561  647  671  729  771  798   822   842   845   879
#> [13,]   17    1  663  671  688  697  705  774  794   810   853   898   943
#> [14,]   17    1  581  634  659  716  733  791  889   916   943   981  1006
#> [15,]   17    1  609  615  646  655  694  749  765   817   851   939   946
#> [16,]   17    1  792  831  850  866  894  951 1001  1019  1021  1026  1048
#> [17,]   17    1  221  575  663  746  843  863  878   904   943   969   974
#> [18,]   17    1  557  583  810  848  908  934  937   962   991  1028  1034
#> [19,]   17    1  271  561  566  657  706  777  851   875   924   939   950
#> [20,]   17    1  556  566  598  654  832  891  907   908   932   947   949
#> [21,]   17    1  221  271  590  777  841  872  889   908   930   934   971
#> [22,]   17    1   20  634  749  753  790  841  857   890   935   987  1010
#> [23,]   17    1  678  694  722  842  844  858  878   883   897   911   943
#> [24,]   17    1  697  729  878  889  904  915  916   920   930   941   947
#> [25,]   16    1  575  590  722  733  784  849  860   882   936  1011  1012
#> [26,]   17    1  656  697  852  883  901  904  920   931   938   966   986
#> [27,]   17    1  488  590  672  830  848  857  858   912   914   931   936
#> [28,]   17    1  488  706  716  733  792  813  830   878   888   899   944
#> [29,]   17    1  645  729  753  857  887  889  894   905   923   937  1001
#> [30,]   17    1  415  717  727  849  868  911  920   921   949   957   961
#> [31,]   15    1   60  561  609  621  753  867  881   893   903   930   949
#> [32,]   17    1  655  671  798  837  840  856  894   897   914   966   973
#> [33,]   13    1  556  589  649  663  755  761  862   908   918   985  1033
#> [34,]   17    1  681  688  728  733  771  826  858   897   907   974   979
#> [35,]   17    1  557  741  753  763  825  853  971   983   997  1003  1010
#> [36,]   17    1   60  656  688  713  717  785  826   850   886   889   897
#> [37,]   17    1  415  562  571  663  680  746  748   794   817   837   863
#> [38,]   17    1  596  623  777  839  900  932  934   939   957  1002  1015
#> [39,]   17    1  791  792  857  867  880  893  896   929   976   977   983
#> [40,]   17    1  562  571  716  828  872  875  887   888   924   957  1001
#> [41,]   14    1  575  687  828  871  882  898  930   946   954   962   999
#> [42,]   17    1  596  621  698  727  746  785  813   831   893   896   912
#> [43,]   17    1  628  649  727  832  833  895  921   923   931   934   949
#> [44,]   17    1   51  596  645  646  694  704  824   852   853   901   933
#> [45,]   16    1   20  741  810  839  852  890  899   907   916   920   967
#> [46,]   17    1  271  566  609  666  763  853  916   989   990   997  1010
#> [47,]   17    1  271  415  642  738  765  817  824   881   890   892   901
#> [48,]   17    1  233  308  677  678  826  857  886   933   938   940   945
#> [49,]   17    1   51  226  833  842  843  849  890   915   948   949  1009
#> [50,]   17    1  233  645  681  688  697  708  753   813   861   875   911
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   883   973   995  1003  1036  1093  1096
#>  [2,]  1052  1053  1057  1072  1081  1091  1096
#>  [3,]   975  1018  1028  1076  1089  1095  1096
#>  [4,]   969   983  1003  1049  1060  1081  1096
#>  [5,]  1027  1033  1042  1080  1087  1095  1096
#>  [6,]   979  1007  1058  1074  1081  1096     0
#>  [7,]   935   941   977   998  1020  1064  1096
#>  [8,]   851   938  1001  1031  1033  1075  1096
#>  [9,]   871   895   901   975   982  1080  1096
#> [10,]   921   937  1020  1021  1096     0     0
#> [11,]  1002  1025  1033  1052  1058  1086  1096
#> [12,]   883   890   914   940  1044  1057  1096
#> [13,]   966   969  1010  1036  1062  1090  1096
#> [14,]  1037  1047  1058  1076  1083  1084  1096
#> [15,]   951   981   984   991   997  1053  1096
#> [16,]  1050  1058  1059  1062  1076  1095  1096
#> [17,]  1007  1012  1029  1043  1047  1087  1096
#> [18,]  1049  1058  1083  1087  1089  1092  1096
#> [19,]   967  1010  1013  1027  1040  1066  1096
#> [20,]   953  1006  1013  1024  1069  1077  1096
#> [21,]   993  1005  1006  1019  1036  1084  1096
#> [22,]  1014  1018  1029  1059  1078  1083  1096
#> [23,]   978  1013  1025  1046  1049  1058  1096
#> [24,]   973   976   994  1007  1039  1075  1096
#> [25,]  1051  1055  1072  1082  1086  1096     0
#> [26,]  1000  1006  1021  1026  1052  1083  1096
#> [27,]   988   989  1025  1038  1045  1084  1096
#> [28,]   945   998  1019  1040  1044  1050  1096
#> [29,]  1018  1031  1034  1040  1070  1077  1096
#> [30,]   977  1000  1011  1020  1027  1047  1096
#> [31,]   950   977  1003  1027  1096     0     0
#> [32,]   985   990  1007  1036  1070  1087  1096
#> [33,]  1034  1049  1096     0     0     0     0
#> [34,]   983   997  1004  1042  1059  1072  1096
#> [35,]  1013  1022  1032  1052  1081  1083  1096
#> [36,]   970   995  1005  1037  1045  1057  1096
#> [37,]   905   932   967   981  1038  1071  1096
#> [38,]  1048  1060  1067  1071  1088  1092  1096
#> [39,]  1018  1021  1026  1073  1083  1091  1096
#> [40,]  1005  1019  1038  1074  1078  1082  1096
#> [41,]  1014  1054  1088  1096     0     0     0
#> [42,]   950   964  1053  1055  1065  1072  1096
#> [43,]   970   989  1015  1020  1039  1064  1096
#> [44,]   936   981  1007  1058  1092  1095  1096
#> [45,]   985   995  1011  1066  1078  1096     0
#> [46,]  1013  1022  1023  1046  1049  1089  1096
#> [47,]   910   972  1000  1036  1056  1092  1096
#> [48,]   974   978  1028  1045  1049  1085  1096
#> [49,]  1021  1039  1042  1052  1071  1079  1096
#> [50,]   976   999  1007  1027  1059  1094  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986