Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_m+1; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   10    1  615  657  678  839  843  865  963   977   986  1003  1096
#>  [2,]    9    1  488  619  704  840  848  878  898  1010  1086  1096     0
#>  [3,]   10    1  614  671  713  722  864  887  897   912   977  1080  1096
#>  [4,]    9    1  590  803  833  932  946  947  948   962  1065  1096     0
#>  [5,]    8    1  415  557  628  817  900  931  957  1079  1096     0     0
#>  [6,]    8    1  415  605  871  877  889  988 1004  1027  1096     0     0
#>  [7,]   10    1  415  615  623  841  857  867  880   955   959  1074  1096
#>  [8,]    5    1  771  839  841  851  994 1096    0     0     0     0     0
#>  [9,]    8    1  415  708  840  848  849  910  978  1005  1096     0     0
#> [10,]   12    1  633  681  774  791  867  978 1010  1019  1024  1071  1073
#> [11,]    8    1  713  836  919  939  991  999 1052  1055  1096     0     0
#> [12,]    8    1  609  780  785  908  975 1024 1058  1069  1096     0     0
#> [13,]    7    1  790  878  888  947  963 1065 1086  1096     0     0     0
#> [14,]    9    1  614  649  671  713  755  824  920   972  1053  1096     0
#> [15,]    6    1  605  629  852  912 1014 1072 1096     0     0     0     0
#> [16,]   10    1  649  659  790  798  945  949  983   985  1010  1072  1096
#> [17,]    6    1  415  666  830  919 1020 1086 1096     0     0     0     0
#> [18,]   10    1  557  571  604  771  846  871  875   889   897  1010  1096
#> [19,]   11    1  415  663  803  823  831  837  931   932   961   999  1080
#> [20,]   15    1  562  598  614  621  843  846  861   893   897   912   926
#> [21,]    6    1  553  571  605  810  877  888 1096     0     0     0     0
#> [22,]    8    1  618  633  746  753  921 1010 1012  1028  1096     0     0
#> [23,]   12    1  618  623  687  761  817  845  921   953   959   980  1047
#> [24,]   11    1  614  768  918  957  990  994 1047  1048  1065  1073  1086
#> [25,]    6    1  642  771  794  859  890  954 1096     0     0     0     0
#> [26,]    6    1  614  659  708  924  975 1078 1096     0     0     0     0
#> [27,]    5    1  672  765  921  988 1029 1096    0     0     0     0     0
#> [28,]   14    1  668  790  794  844  863  887  901   918  1023  1032  1065
#> [29,]   12    1  488  838  845  852  863  870  904   948   951   982  1036
#> [30,]    6    1  596  935  959  961 1004 1019 1096     0     0     0     0
#> [31,]    8    1  619  856  861  894  986  999 1005  1045  1096     0     0
#> [32,]   10    1  728  785  809  859  890  953  970  1005  1037  1072  1096
#> [33,]    8    1  556  575  598  746  868  934 1003  1012  1096     0     0
#> [34,]   11    1  561  748  822  828  850  861  932   944   998  1048  1065
#> [35,]    5    1  657  753  830 1002 1086 1096    0     0     0     0     0
#> [36,]   13    1  415  553  575  628  741  748  837   877   935   962   978
#> [37,]   10    1  575  598  629  877  883  903  938   949   990  1024  1096
#> [38,]    7    1  561  765  860  967 1017 1051 1074  1096     0     0     0
#> [39,]    9    1  589  623  642  748  824  862  918  1013  1046  1096     0
#> [40,]    6    1  614  708  841 1035 1039 1075 1096     0     0     0     0
#> [41,]   14    1  415  566  666  753  830  844  857   867   870   888   924
#> [42,]    8    1  681  841  846  940  980 1008 1039  1080  1096     0     0
#> [43,]    9    1  415  615  633  817  828  860  875   937  1062  1096     0
#> [44,]   12    1  581  634  708  846  863  901  903   930   953  1039  1067
#> [45,]    8    1  415  678  824  848  898 1019 1038  1086  1096     0     0
#> [46,]    9    1  798  850  924  946 1011 1020 1031  1068  1076  1096     0
#> [47,]    8    1  415  748  858  923  932  936  955  1021  1096     0     0
#> [48,]    8    1  556  825  837  861  976  999 1048  1084  1096     0     0
#> [49,]   14    1  562  657  746  794  839  842  892   907   910  1014  1058
#> [50,]    8    1  598  619  634  920  926  974  990  1009  1096     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]     0     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]  1086  1096     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]  1096     0     0     0     0     0     0
#> [20,]  1000  1057  1074  1076  1096     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]  1082  1096     0     0     0     0     0
#> [24,]  1096     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]  1067  1076  1086  1096     0     0     0
#> [29,]  1068  1096     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]  1096     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]  1079  1086  1096     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]   994  1018  1056  1096     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0
#> [44,]  1086  1096     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]  1064  1073  1086  1096     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1   51  694  705  733  768  790  845  883  888  890  929  930  977
#> [16]  994  997 1069 1072 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   52   78  461  514  810  836  849  866  873  903  920  939  996
#> [16]  999 1002 1011 1090 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  415  431  525  657  683  740  816  863  873  887  899  912  933
#> [16] 1023 1050 1054 1056 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1  261  686  731  780  781  796  836  850  930  932  944  950  960
#> [16]  983 1000 1028 1057 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   15    1  646  687  768  841  848  877  931   944   947   972  1000
#>  [2,]   17    1  221  581  619  646  649  777  785   831   924   962   966
#>  [3,]   15    1  221  634  680  785  831  854  875   935   967   997  1010
#>  [4,]   13    1  657  711  851  915  953  983  992  1017  1018  1024  1049
#>  [5,]   17    1  566  628  656  738  792  813  867   887   888   930   963
#>  [6,]   17    1  354  605  645  838  859  904  946   954   962   970   996
#>  [7,]   17    1  619  678  694  741  785  848  867   885   908   953   955
#>  [8,]   17    1  655  659  784  843  860  908  936   947   959   982  1011
#>  [9,]   17    1  226  548  613  792  817  826  857   867   915   937   953
#> [10,]   17    1  308  556  557  614  634  713  822   840   872   875   880
#> [11,]   16    1   20  557  581  583  711  713  753   780   836   894   912
#> [12,]   14    1   60  556  678  771  777  822  887   977  1011  1016  1020
#> [13,]   17    1   60  604  663  733  749  840  842   843   845   850   871
#> [14,]   16    1  598  604  694  755  765  825  843   866   884   937   940
#> [15,]   17    1  548  613  642  646  663  698  824   825   836   843   881
#> [16,]   17    1  308  623  656  666  681  697  725   771   826   852   888
#> [17,]   13    1  557  590  704  719  780  810  824   858   876   920   966
#> [18,]   17    1   51  657  671  741  761  784  833   842   905   911   937
#> [19,]   17    1  656  725  761  774  791  825  841   907   931   934   963
#> [20,]   17    1  629  654  753  768  825  845  883   891   898   944   953
#> [21,]   17    1  556  671  705  774  813  870  924   968   969   978   979
#> [22,]   17    1  233  308  623  672  755  785  838   858   882   889   914
#> [23,]   17    1   51  633  655  687  708  725  803   809   968  1002  1010
#> [24,]   17    1  581  659  672  705  738  765  854   871   883   895   924
#> [25,]   17    1  583  623  668  671  728  771  817   849   856   888   894
#> [26,]   17    1  271  613  677  687  774  780  813   838   888   889   907
#> [27,]   17    1   20  233  271  859  866  878  893   914   928   930   932
#> [28,]   13    1  557  613  717  774  823  824  830   857   919   937   945
#> [29,]   14    1  596  705  716  828  841  852  947   962   985   990  1031
#> [30,]   17    1  727  824  850  882  888  889  946   953   986   992   995
#> [31,]   17    1  233  629  722  817  837  844  851   888   898   947   963
#> [32,]   17    1   20  659  817  845  853  861  872   875   908   916   941
#> [33,]   17    1  226  415  575  633  666  678  791   792   861   912   975
#> [34,]   17    1  221  629  668  722  828  875  890   903   941   964  1021
#> [35,]   17    1  233  615  825  852  857  883  893   900   918   963   973
#> [36,]   17    1  553  618  671  729  858  881  929   932   935   936   976
#> [37,]   17    1  688  717  722  780  839  850  856   870   881   886   897
#> [38,]   17    1  221  566  581  615  755  852  865   897   946   995  1001
#> [39,]   16    1  604  655  657  706  822  834  840   841   944   961   996
#> [40,]   17    1   20  609  681  791  836  877  893   905   908   912   916
#> [41,]   17    1  557  763  784  824  850  861  862   901   904   923   932
#> [42,]   17    1  596  619  687  716  727  729  810   848   872   895   921
#> [43,]   16    1   20  566  678  708  828  837  853   865   919   941   991
#> [44,]   17    1  575  596  647  697  705  794  844   871   882   943   970
#> [45,]   17    1  598  609  645  753  841  875  878   973   974   993  1017
#> [46,]   17    1  571  614  618  694  748  777  794   832   849   871   877
#> [47,]   17    1  613  647  755  824  843  896  899   915   944   948   967
#> [48,]   17    1  488  618  628  659  716  864  870   875   877   912   948
#> [49,]   17    1  221  226  659  728  836  877  895   949   959   973  1024
#> [50,]   17    1  415  562  615  633  649  666  677   694   755   839   880
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  1033  1059  1065  1095  1096     0     0
#>  [2,]   983  1011  1023  1039  1048  1074  1096
#>  [3,]  1025  1028  1060  1065  1096     0     0
#>  [4,]  1079  1094  1096     0     0     0     0
#>  [5,]   964   999  1021  1025  1086  1093  1096
#>  [6,]  1028  1032  1046  1053  1085  1095  1096
#>  [7,]   964   972  1016  1037  1085  1091  1096
#>  [8,]  1035  1038  1052  1066  1086  1087  1096
#>  [9,]   972   974  1029  1072  1084  1094  1096
#> [10,]   881   908   962   972  1008  1077  1096
#> [11,]   950   966   993  1005  1018  1096     0
#> [12,]  1028  1053  1095  1096     0     0     0
#> [13,]   885   907   947  1046  1062  1080  1096
#> [14,]   943   979   981  1038  1067  1096     0
#> [15,]   886  1011  1021  1035  1060  1069  1096
#> [16,]   933   945   971  1005  1030  1053  1096
#> [17,]   994  1013  1096     0     0     0     0
#> [18,]   940   995  1061  1062  1077  1093  1096
#> [19,]   969  1038  1048  1067  1070  1084  1096
#> [20,]  1010  1020  1050  1060  1068  1090  1096
#> [21,]   983   994  1008  1037  1062  1066  1096
#> [22,]   918   991   994  1002  1013  1022  1096
#> [23,]  1030  1049  1060  1064  1083  1094  1096
#> [24,]   929   973   977  1028  1041  1076  1096
#> [25,]   931   963   988   999  1046  1077  1096
#> [26,]   910   924   989  1018  1036  1085  1096
#> [27,]   962  1005  1013  1044  1074  1078  1096
#> [28,]   963  1014  1096     0     0     0     0
#> [29,]  1032  1038  1064  1096     0     0     0
#> [30,]  1014  1020  1035  1044  1082  1083  1096
#> [31,]   967   970  1018  1064  1074  1089  1096
#> [32,]   962   971   994  1002  1038  1055  1096
#> [33,]   989  1022  1029  1051  1071  1072  1096
#> [34,]  1028  1034  1035  1059  1061  1091  1096
#> [35,]   980  1016  1018  1046  1055  1093  1096
#> [36,]   993   999  1015  1050  1088  1095  1096
#> [37,]   934   938  1031  1069  1083  1091  1096
#> [38,]  1005  1006  1023  1036  1078  1087  1096
#> [39,]  1029  1044  1060  1072  1092  1096     0
#> [40,]   924   989  1012  1020  1024  1038  1096
#> [41,]   935   937   993  1000  1039  1060  1096
#> [42,]   932   946  1013  1023  1025  1079  1096
#> [43,]   994  1007  1012  1041  1085  1096     0
#> [44,]   974  1015  1050  1051  1068  1079  1096
#> [45,]  1026  1034  1048  1051  1062  1087  1096
#> [46,]   881   884   908   947   954  1057  1096
#> [47,]   987  1018  1022  1027  1052  1068  1096
#> [48,]   981   987  1043  1046  1056  1088  1096
#> [49,]  1025  1027  1040  1045  1046  1053  1096
#> [50,]   889  1010  1029  1037  1039  1064  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986