Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_m+1; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    8    1  831  885  955  957  989 1012 1023  1086  1096     0     0
#>  [2,]    8    1  711  722  741  852  968  991 1083  1085  1096     0     0
#>  [3,]    7    1  706  817  896  993 1025 1041 1086  1096     0     0     0
#>  [4,]   12    1  598  657  722  822  839  886  898   910   987   994  1068
#>  [5,]   12    1  553  598  646  828  962  974  986   990  1001  1013  1069
#>  [6,]    9    1  561  575  828  868  886  934 1040  1045  1061  1096     0
#>  [7,]   11    1  719  753  822  830  881  941  972   978   992   998  1060
#>  [8,]   11    1  727  830  834  839  876  892  904   937   960   986  1069
#>  [9,]   11    1  649  705  711  722  822  832  919   993  1007  1064  1086
#> [10,]    6    1  749  813  888  945  997 1058 1096     0     0     0     0
#> [11,]   13    1  621  633  642  722  791  901  949   988  1007  1039  1043
#> [12,]    6    1  556  823  962  993 1008 1071 1096     0     0     0     0
#> [13,]   11    1  681  803  828  833  856  881  911   929   951   957   971
#> [14,]   17    1  557  598  678  719  729  741  777   794   809   810   831
#> [15,]    7    1  741  763  840  937 1029 1077 1086  1096     0     0     0
#> [16,]    4    1  415  634  982 1077 1096    0    0     0     0     0     0
#> [17,]    8    1  784  794  883  924  992 1041 1042  1053  1096     0     0
#> [18,]    9    1  647  654  832  868  870  955  982   987  1084  1096     0
#> [19,]    9    1  415  741  810  959  961  963  992  1002  1029  1096     0
#> [20,]    8    1  556  613  843  853  894  943  979  1041  1096     0     0
#> [21,]   10    1  566  771  833  851  876  957 1004  1061  1065  1079  1096
#> [22,]   11    1  609  680  738  780  843  895  972  1015  1041  1046  1075
#> [23,]   11    1  590  854  903  921  944  964  971   977  1020  1036  1086
#> [24,]   10    1  666  719  741  784  833  836  867   957  1007  1066  1096
#> [25,]    6    1  850  911  941 1001 1014 1045 1096     0     0     0     0
#> [26,]    9    1  618  619  698  853  903  908  973  1052  1086  1096     0
#> [27,]   10    1  642  719  790  889  933 1006 1054  1055  1058  1062  1096
#> [28,]   10    1  581  645  719  834  880 1009 1016  1065  1066  1070  1096
#> [29,]   10    1  415  605  629  823  825  826  926  1026  1041  1085  1096
#> [30,]   14    1  647  677  713  794  830  831  870   934   961  1011  1036
#> [31,]    6    1  619  719  900  946 1050 1054 1096     0     0     0     0
#> [32,]   12    1  553  823  870  895  900  938  992   993  1067  1072  1073
#> [33,]   11    1  415  548  826  830  857  865  884   950   960  1009  1063
#> [34,]    4    1  553  964 1065 1073 1096    0    0     0     0     0     0
#> [35,]    4    1  621  681  868  890 1096    0    0     0     0     0     0
#> [36,]    8    1  697  722  753  881  930  982 1046  1049  1096     0     0
#> [37,]    7    1  557  623  680  708  838  871 1044  1096     0     0     0
#> [38,]   11    1  671  777  832  854  896  908  969  1031  1041  1052  1086
#> [39,]   10    1  581  733  867  885  899  935  941   981   999  1051  1096
#> [40,]    9    1  548  556  672  727  879  883  934   936   959  1096     0
#> [41,]    6    1  790  794  861  951  959  986 1096     0     0     0     0
#> [42,]    8    1  623  697  777  901  926 1002 1033  1051  1096     0     0
#> [43,]   13    1  659  697  706  719  826  856  929   932   964  1032  1053
#> [44,]    8    1  553  896  926  937 1005 1035 1072  1080  1096     0     0
#> [45,]   10    1  784  875  897  930  936  968  989  1001  1037  1053  1096
#> [46,]   10    1  583  749  817  868  941  955 1015  1021  1032  1048  1096
#> [47,]    7    1  654  896  938  994 1000 1003 1079  1096     0     0     0
#> [48,]    8    1  659  663  741  798  886  986  995  1000  1096     0     0
#> [49,]   16    1  589  614  649  678  741  832  842   911   932   944   955
#> [50,]   11    1  575  659  697  698  777  822  924   926   948  1061  1086
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]     0     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]  1072  1096     0     0     0     0     0
#>  [5,]  1083  1096     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]  1096     0     0     0     0     0     0
#>  [8,]  1096     0     0     0     0     0     0
#>  [9,]  1096     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]  1053  1069  1096     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]  1096     0     0     0     0     0     0
#> [14,]   837   844   898   901   938   969  1096
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]  1096     0     0     0     0     0     0
#> [23,]  1096     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]  1047  1053  1060  1096     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]  1086  1096     0     0     0     0     0
#> [33,]  1096     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]  1096     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]  1055  1081  1096     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]   964   982   988   989  1022  1096     0
#> [50,]  1096     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  233  354  553  634  698  733  862  865  886  890  919  940  970
#> [16]  971  988  992 1020 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   16   56  168  263  300  407  493  560  593  651  723  781  805
#> [16]  941  955 1005 1014 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  382  408  464  483  525  529  744  786  793  830  855  873  917
#> [16]  991 1004 1056 1095 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1  554  686  750  808  832  839  846  866  912  923  943  948  963
#> [16]  975 1018 1036 1060 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   16    1  556  615  642  672  716  728  761   823   824   830   848
#>  [2,]   17    1  659  722  853  870  905  910  921   943   954   969   981
#>  [3,]   17    1  575  605  656  706  833  881  892   930   940   946   959
#>  [4,]   17    1  609  680  741  763  825  833  864   951   954   980   989
#>  [5,]   17    1  590  618  619  642  705  716  719   755   785   843   887
#>  [6,]   17    1   20   51  609  725  728  790  836   894   900   924   945
#>  [7,]   17    1  557  566  848  853  861  880  893   962   971   975   986
#>  [8,]   17    1  556  583  589  613  746  864  865   905   918   953  1035
#>  [9,]   17    1  415  671  774  834  842  908  914   943   954   979   981
#> [10,]   17    1  233  657  824  854  868  877  894   899   907   910   950
#> [11,]   17    1   60  671  677  763  837  853  878   928   932   946   992
#> [12,]   16    1  798  851  860  870  888  892  893   932   961   993   996
#> [13,]   17    1  596  609  633  654  768  785  861   870   877   894   963
#> [14,]   17    1  415  590  623  713  727  761  785   832   898   901   916
#> [15,]   17    1   60  680  694  698  830  858  859   875   876   903   931
#> [16,]   17    1  562  604  646  657  694  784  798   822   837   842   875
#> [17,]   14    1  271  642  645  678  780  839  851   868   870   892   932
#> [18,]   17    1  708  791  794  813  824  833  848   863   930   943   978
#> [19,]   16    1  354  614  628  656  704  711  822   833   836   868   885
#> [20,]   17    1  354  556  777  823  836  852  863   881   885   968   988
#> [21,]   17    1  226  583  604  623  642  728  733   810   839   842   879
#> [22,]   17    1  634  645  666  678  706  774  845   876   923   932   937
#> [23,]   17    1  561  671  687  719  738  852  860   878   935   936   954
#> [24,]   17    1  654  657  780  831  879  912  914   921   929   970   990
#> [25,]   17    1  221  561  614  645  657  666  688   813   860   977  1000
#> [26,]   17    1  604  704  803  834  838  843  850   859   865   880   882
#> [27,]   15    1  415  623  629  677  840  866  891   899   934   946   964
#> [28,]   17    1  581  596  663  716  727  777  859   863   911   919   936
#> [29,]   17    1  590  657  774  785  803  813  839   850   889   934   950
#> [30,]   13    1  221  561  613  697  704  862  866   979  1009  1010  1020
#> [31,]   17    1  354  687  791  803  872  889  905   915   924  1000  1018
#> [32,]   17    1  609  678  681  719  768  780  784   798   862   866   878
#> [33,]   17    1  581  604  605  623  698  729  813   828   844   878   901
#> [34,]   17    1  605  618  656  678  680  794  817   870   878   899   977
#> [35,]   17    1  556  646  678  680  761  846  867   882   899   935   984
#> [36,]   17    1  589  604  609  614  649  704  713   768   828   851   905
#> [37,]   17    1  583  598  621  698  833  838  858   894   916   970   984
#> [38,]   17    1   20  308  571  633  654  708  833   849   860   911   929
#> [39,]   17    1  647  705  791  803  833  837  859   871   941   945   948
#> [40,]   17    1  621  628  659  755  813  846  878   916   923   933   960
#> [41,]   17    1  589  655  666  705  706  784  834   843   884   903   936
#> [42,]   17    1  581  668  755  825  832  872  912   954  1011  1027  1042
#> [43,]   17    1  618  663  666  761  831  852  904   931   933   936   955
#> [44,]   17    1   20  556  590  642  755  777  794   833   851   889   955
#> [45,]   16    1   20  583  656  746  755  824  839   865   870   912   928
#> [46,]   17    1  271  619  663  733  825  830  844   851   876   943   978
#> [47,]   17    1  221  613  615  647  842  904  948   951   953   954   960
#> [48,]   17    1  415  583  708  716  832  872  875   876   877   890   912
#> [49,]   17    1  271  604  728  824  856  939  944   946   963   990  1003
#> [50,]   17    1  556  557  654  681  688  727  813   834   848   876   877
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  1023  1045  1077  1081  1082  1096     0
#>  [2,]   989   993  1011  1070  1075  1089  1096
#>  [3,]   991   996  1015  1034  1046  1083  1096
#>  [4,]  1000  1004  1036  1040  1061  1074  1096
#>  [5,]   901  1016  1021  1066  1075  1089  1096
#>  [6,]   969  1032  1037  1052  1073  1079  1096
#>  [7,]   988   991  1002  1040  1052  1095  1096
#>  [8,]  1060  1072  1073  1076  1085  1095  1096
#>  [9,]   995  1012  1041  1063  1075  1077  1096
#> [10,]   955   979   996  1018  1048  1083  1096
#> [11,]   998  1015  1019  1048  1050  1081  1096
#> [12,]  1033  1039  1041  1058  1083  1096     0
#> [13,]   970   977   979   992  1060  1094  1096
#> [14,]   921   983   986   996  1049  1068  1096
#> [15,]   946   962   989   996  1001  1075  1096
#> [16,]   959   975  1010  1016  1046  1086  1096
#> [17,]   966   992  1035  1096     0     0     0
#> [18,]   979   997  1003  1058  1091  1094  1096
#> [19,]   896   945   959   982  1051  1096     0
#> [20,]   998  1020  1038  1064  1077  1095  1096
#> [21,]   920  1007  1013  1019  1061  1067  1096
#> [22,]   954   979  1029  1041  1050  1091  1096
#> [23,]   973   991  1002  1006  1034  1078  1096
#> [24,]  1026  1047  1049  1071  1087  1093  1096
#> [25,]  1002  1003  1015  1026  1046  1065  1096
#> [26,]   926   954   961  1002  1025  1027  1096
#> [27,]   985  1044  1066  1078  1096     0     0
#> [28,]   937   959  1016  1021  1031  1069  1096
#> [29,]   981   987  1004  1031  1046  1084  1096
#> [30,]  1033  1043  1096     0     0     0     0
#> [31,]  1029  1043  1073  1077  1081  1087  1096
#> [32,]   898   985   996  1026  1028  1038  1096
#> [33,]   955   990  1004  1064  1084  1094  1096
#> [34,]  1023  1030  1036  1040  1073  1090  1096
#> [35,]  1005  1014  1028  1042  1051  1057  1096
#> [36,]   920   969  1003  1036  1077  1080  1096
#> [37,]   990  1002  1004  1013  1020  1029  1096
#> [38,]   945   970   978  1006  1042  1048  1096
#> [39,]   993  1026  1031  1049  1068  1090  1096
#> [40,]   984   991  1061  1064  1088  1092  1096
#> [41,]   966   969   976  1000  1012  1024  1096
#> [42,]  1058  1060  1064  1076  1089  1090  1096
#> [43,]   989  1001  1004  1027  1060  1079  1096
#> [44,]   972   979  1026  1032  1040  1050  1096
#> [45,]  1001  1007  1040  1049  1079  1096     0
#> [46,]   986   991  1023  1024  1041  1060  1096
#> [47,]   976   984   999  1052  1078  1085  1096
#> [48,]   919   939   996  1026  1058  1093  1096
#> [49,]  1005  1034  1051  1071  1075  1092  1096
#> [50,]   928  1009  1033  1035  1040  1064  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986