Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   11    1  609  659  694  823  854  880  930   984  1016  1051  1064
#>  [2,]    7    1  698  875  881  891  989 1027 1039  1096     0     0     0
#>  [3,]    7    1  598  677  868  883  910  992 1016  1096     0     0     0
#>  [4,]   10    1  634  708  711  763  894  899  941   949  1020  1025  1096
#>  [5,]   11    1  598  621  698  785  866  980  992  1007  1011  1015  1038
#>  [6,]    9    1  678  716  777  837  895  926  972  1015  1074  1096     0
#>  [7,]    7    1  790  822  832  949  984 1024 1051  1096     0     0     0
#>  [8,]    6    1  562  677  853  854  857 1032 1096     0     0     0     0
#>  [9,]    8    1  704  719  738  875  945  993 1000  1055  1096     0     0
#> [10,]    6    1  668  733  888  941 1033 1035 1096     0     0     0     0
#> [11,]    9    1  415  681  753  771  817  905  937  1000  1026  1096     0
#> [12,]   13    1  553  557  575  629  688  698  771   915  1004  1010  1041
#> [13,]   10    1  488  589  824  930  972  981 1000  1013  1065  1077  1096
#> [14,]   12    1  556  681  704  813  830  871  891   990  1013  1028  1048
#> [15,]    9    1  785  856  890  946  961  963 1052  1072  1078  1096     0
#> [16,]    5    1  628  941 1041 1081 1082 1096    0     0     0     0     0
#> [17,]   12    1  557  575  581  886  962  963 1001  1013  1016  1018  1042
#> [18,]    8    1  557  571  842  879  926 1064 1077  1086  1096     0     0
#> [19,]    9    1  553  556  614  854  898  914  924   966  1079  1096     0
#> [20,]    9    1  605  698  803  905  970  972  993  1018  1055  1096     0
#> [21,]    8    1  598  677  844  892  964  968 1014  1024  1096     0     0
#> [22,]    9    1  415  663  879  896  905  934  946  1007  1079  1096     0
#> [23,]    8    1  604  889  899  932 1005 1044 1085  1086  1096     0     0
#> [24,]    8    1  614  628  655  722  889  923  961  1033  1096     0     0
#> [25,]    9    1  415  604  706  824  832  892 1044  1057  1079  1096     0
#> [26,]    8    1  561  604  705  846  897  960  993  1077  1096     0     0
#> [27,]   10    1  634  668  817  888  928  982 1013  1031  1044  1078  1096
#> [28,]   12    1  415  605  687  719  790  831  885   891   948   967  1019
#> [29,]    5    1  924  945 1004 1016 1018 1096    0     0     0     0     0
#> [30,]   10    1  822  843  845  878  950  959  984  1003  1015  1044  1096
#> [31,]    7    1  647  677  971  994 1074 1077 1086  1096     0     0     0
#> [32,]    8    1  623  678  733  889  968 1014 1028  1086  1096     0     0
#> [33,]    8    1  561  646  844  966 1010 1018 1024  1079  1096     0     0
#> [34,]    5    1  604  784  914  945  973 1096    0     0     0     0     0
#> [35,]   10    1  415  688  722  763  824  831  893   945  1015  1080  1096
#> [36,]   11    1  575  613  668  727  842  846  871   887   929  1002  1077
#> [37,]    9    1  571  621  729  837  840 1010 1030  1048  1076  1096     0
#> [38,]    9    1  605  688  825  893  915  948 1065  1076  1078  1096     0
#> [39,]   12    1  415  589  746  834  842  918  959   966   976   981  1069
#> [40,]    6    1  415  623  663  953 1050 1063 1096     0     0     0     0
#> [41,]    5    1  562  583  970 1078 1085 1096    0     0     0     0     0
#> [42,]   15    1  604  719  722  765  872  886  888   889   893   916   949
#> [43,]   11    1  680  763  771  836  880  893  976   981  1004  1020  1044
#> [44,]   10    1  645  824  886  900  910  963 1003  1005  1023  1086  1096
#> [45,]    9    1  713  719  834  881  943  963 1070  1077  1084  1096     0
#> [46,]   10    1  633  877  932  955  978 1003 1016  1042  1043  1072  1096
#> [47,]    8    1  678  728  846  887  953  973  984   997  1096     0     0
#> [48,]    7    1  898  966  973  997 1013 1072 1080  1096     0     0     0
#> [49,]    8    1  556  619  803  832  853  897  946   982  1096     0     0
#> [50,]    7    1  553  633  858  882  900  960 1081  1096     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  1096     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]  1096     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0
#> [12,]  1056  1080  1096     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]  1086  1096     0     0     0     0     0
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]  1049  1096     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]     0     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]  1051  1096     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]  1096     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0
#> [39,]  1086  1096     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]  1021  1023  1076  1083  1096     0     0
#> [43,]  1096     0     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  553  642  697  749  817  846  856  882  912  914  926  960  976
#> [16]  982 1049 1072 1084 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   39  380  405  636  664  777  834  845  885  916  924  936 1035
#> [16] 1069 1072 1083 1090 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  377  439  451  525  651  765  787  795  832  848  937  947  962
#> [16]  984 1051 1070 1089 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   15    1  191  292  693  783  933  977  987  993 1003 1004 1038 1042 1057
#> [16] 1064 1087 1096    0    0

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  354  619  708  717  834  881  887   953   967   970   976
#>  [2,]   17    1   20  562  634  688  697  876  877   894   908   939   953
#>  [3,]   17    1  308  590  614  657  678  694  716   803   832   852   858
#>  [4,]   17    1  221  562  618  642  722  833  846   860   887   896   905
#>  [5,]   17    1  680  831  849  900  940  951  957   964   974   983   984
#>  [6,]   10    1  354  791  882  893 1013 1014 1029  1041  1054  1091  1096
#>  [7,]   17    1   51  415  666  719  753  774  856   865   928   939   966
#>  [8,]   16    1  488  668  729  817  834  838  851   898   915   916   938
#>  [9,]   17    1  556  557  609  656  687  727  791   851   895   912   951
#> [10,]   17    1  553  598  634  680  838  842  852   853   858   908   915
#> [11,]   17    1  615  628  688  722  761  844  860   871   875   919   933
#> [12,]   17    1   20  233  672  774  840  889  896   931   936   943   949
#> [13,]   17    1  688  722  863  887  891  899  930   936   938   962   986
#> [14,]   17    1   51  575  590  681  708  813  833   838   928   963   964
#> [15,]   17    1  271  488  571  619  621  629  649   728   780   854   890
#> [16,]   17    1   20  645  647  655  677  698  711   727   738   886   935
#> [17,]   17    1  415  575  589  768  774  833  846   858   938   947   966
#> [18,]   17    1  614  677  741  748  792  830  863   871   895   928   979
#> [19,]   17    1  557  621  688  761  803  809  848   856   899   933   954
#> [20,]   17    1  415  589  774  831  892  908  950   951   964   969   970
#> [21,]   16    1  488  642  771  832  834  867  879   931   938   984   985
#> [22,]   17    1  697  719  748  753  832  850  851   857   883   900   907
#> [23,]   14    1   51  557  654  717  729  823  850   851   858   914   921
#> [24,]   15    1  561  581  642  671  728  823  839   866   871   914   994
#> [25,]   17    1   20  628  729  763  826  828  831   846   889   940   960
#> [26,]   17    1  678  680  698  753  774  790  867   877   930   935   943
#> [27,]   17    1  598  618  634  649  706  822  830   849   853   860   867
#> [28,]   17    1   60  233  553  556  557  562  646   647   868   910   934
#> [29,]   17    1  687  733  791  866  881  899  928   944   948   972   974
#> [30,]   17    1   20  221  621  657  706  725  763   834   848   891   897
#> [31,]   17    1  728  771  780  791  834  843  848   852   853   899   903
#> [32,]   17    1  566  589  598  717  780  833  839   846   860   866   961
#> [33,]   13    1  566  619  642  784  859  912  950   954   977  1006  1009
#> [34,]   17    1  817  833  838  839  845  871  883   940   950   954   979
#> [35,]   17    1  575  598  647  705  719  791  828   842   860   894   898
#> [36,]   17    1  581  615  623  649  698  711  748   768   870   926   936
#> [37,]   17    1   20  581  748  761  851  863  880   881   885   921   944
#> [38,]   17    1  553  704  792  857  883  885  920   951   993  1012  1022
#> [39,]   17    1  354  488  561  628  649  680  763   791   817   838   865
#> [40,]   17    1  657  722  748  768  877  884  893   923   924   926   938
#> [41,]   17    1   51  604  605  647  729  733  753   839   864   888   931
#> [42,]   17    1  415  589  609  634  654  784  790   834   859   910   977
#> [43,]   17    1  548  649  822  840  875  897  911   918   937   961   973
#> [44,]   17    1  583  589  604  618  628  649  738   864   872   881   894
#> [45,]   17    1  226  656  688  765  780  784  837   844   887   900   983
#> [46,]   17    1  571  596  656  698  803  860  903   916   962   973   999
#> [47,]   17    1  548  704  725  768  794  813  850   854   872   890   915
#> [48,]   17    1  233  354  668  694  706  733  771   790   822   899   904
#> [49,]   17    1  596  647  719  753  849  879  891   921   966   973   991
#> [50,]   17    1  354  583  598  613  823  833  842   878   916   962   974
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   991   999  1004  1041  1059  1066  1096
#>  [2,]   967   981   992  1040  1072  1087  1096
#>  [3,]   926   933   934   935  1046  1065  1096
#>  [4,]   914   994  1011  1039  1045  1082  1096
#>  [5,]  1004  1030  1037  1049  1053  1074  1096
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]   985  1019  1029  1040  1057  1083  1096
#>  [8,]   969  1005  1036  1057  1067  1096     0
#>  [9,]   953   957  1024  1040  1043  1087  1096
#> [10,]   923   978   991   997  1029  1045  1096
#> [11,]   963   985  1053  1062  1068  1072  1096
#> [12,]   959   981  1017  1019  1026  1084  1096
#> [13,]   991   998  1008  1029  1034  1077  1096
#> [14,]   979   982  1038  1058  1062  1064  1096
#> [15,]   903   910   915   961   978  1014  1096
#> [16,]   937   943   991  1015  1034  1087  1096
#> [17,]   992  1009  1038  1040  1049  1078  1096
#> [18,]   982   995   997  1014  1034  1092  1096
#> [19,]   977   993  1027  1062  1082  1090  1096
#> [20,]   991   994   997   999  1036  1080  1096
#> [21,]  1027  1033  1083  1086  1094  1096     0
#> [22,]   928   955   966   969  1018  1024  1096
#> [23,]  1011  1042  1062  1096     0     0     0
#> [24,]  1000  1015  1060  1083  1096     0     0
#> [25,]   982  1011  1024  1042  1051  1065  1096
#> [26,]   944   970  1022  1023  1036  1093  1096
#> [27,]   882   990  1017  1024  1041  1091  1096
#> [28,]   945   982  1026  1049  1056  1093  1096
#> [29,]   981   996  1040  1078  1089  1091  1096
#> [30,]   904   937   941   995  1037  1068  1096
#> [31,]   940   955   994   995  1008  1039  1096
#> [32,]   969   997  1029  1031  1086  1088  1096
#> [33,]  1083  1087  1096     0     0     0     0
#> [34,]  1007  1045  1052  1054  1058  1081  1096
#> [35,]   904   954   989  1007  1035  1046  1096
#> [36,]   953   966  1024  1037  1056  1090  1096
#> [37,]   959  1000  1071  1077  1078  1091  1096
#> [38,]  1024  1035  1048  1057  1079  1080  1096
#> [39,]   871   890  1007  1016  1055  1088  1096
#> [40,]   954   969  1018  1050  1062  1074  1096
#> [41,]   949   980  1026  1029  1063  1071  1096
#> [42,]  1000  1054  1055  1069  1075  1076  1096
#> [43,]   983   995   999  1016  1080  1095  1096
#> [44,]   900   915   930   977   990  1015  1096
#> [45,]   987   992  1003  1010  1029  1062  1096
#> [46,]  1003  1018  1041  1075  1076  1079  1096
#> [47,]   935   939   951  1003  1009  1077  1096
#> [48,]   940   970   990   995  1028  1094  1096
#> [49,]  1028  1042  1059  1061  1089  1092  1096
#> [50,]  1008  1026  1028  1034  1047  1061  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986