Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   10    1  415  598  746  857  867  881  923   949  1001  1077  1096
#>  [2,]    7    1  672  846  870  950  953  996 1073  1096     0     0     0
#>  [3,]   11    1  415  562  716  867  886  892  911   928   932  1030  1058
#>  [4,]    7    1  840  878  891  904  932 1012 1060  1096     0     0     0
#>  [5,]    9    1  553  711  748  771  879  951  974  1006  1086  1096     0
#>  [6,]    8    1  561  575  777  791  932  945 1055  1072  1096     0     0
#>  [7,]   10    1  604  719  722  790  851  860  912   945   995  1086  1096
#>  [8,]   10    1  598  833  895  896  979 1005 1013  1042  1067  1077  1096
#>  [9,]   10    1  609  716  825  830  860  914  920   932   992  1053  1096
#> [10,]    9    1  792  889  897  914  961  969  993  1023  1071  1096     0
#> [11,]    9    1  609  785  790  822  824  919  969   995  1054  1096     0
#> [12,]   10    1  553  556  596  901  962  992 1006  1011  1031  1075  1096
#> [13,]    8    1  864  871  941  968  973 1039 1046  1078  1096     0     0
#> [14,]   10    1  556  633  748  785  826  923  946  1044  1077  1080  1096
#> [15,]    6    1  548  589  634  844  889 1074 1096     0     0     0     0
#> [16,]    8    1  581  590  771  967  993 1005 1047  1055  1096     0     0
#> [17,]   13    1  809  839  859  888  941  959  968   969   996  1011  1027
#> [18,]    9    1  844  856  882  907  973 1030 1049  1068  1074  1096     0
#> [19,]    9    1  488  561  728  791  822  866  974   981  1085  1096     0
#> [20,]   10    1  415  614  697  749  825  838  880   930   997  1054  1096
#> [21,]   12    1  556  694  780  852  877  883  891   892   912   943   975
#> [22,]   16    1  659  748  777  780  794  867  868   920   970   975  1002
#> [23,]   13    1  566  657  706  719  840  868  920   931   950   987  1037
#> [24,]    8    1  566  777  846  870  903  924  990  1018  1096     0     0
#> [25,]   12    1  646  780  825  831  931  960  992  1005  1032  1068  1072
#> [26,]   11    1  575  646  871  953  955  974 1007  1042  1055  1057  1082
#> [27,]    9    1  561  845  904  911  926  968  969  1046  1050  1096     0
#> [28,]   10    1  677  838  841  854  894  914  946  1019  1022  1036  1096
#> [29,]    6    1  883  945  996 1030 1062 1083 1096     0     0     0     0
#> [30,]    7    1  839  947  968  987 1006 1023 1050  1096     0     0     0
#> [31,]    5    1  605  785  830  933  944 1096    0     0     0     0     0
#> [32,]    9    1  634  680  823  870  886  954  995   998  1054  1096     0
#> [33,]    6    1  646  824  877  920  924 1051 1096     0     0     0     0
#> [34,]   12    1  618  785  884  908  919  945  987   995  1030  1052  1060
#> [35,]    7    1  856  867  990  997 1028 1036 1076  1096     0     0     0
#> [36,]   10    1  561  605  687  717  871  912  920   994  1044  1086  1096
#> [37,]    6    1  613  618  738  836 1008 1057 1096     0     0     0     0
#> [38,]   10    1  654  708  946  948  971  976 1012  1053  1056  1086  1096
#> [39,]    8    1  415  890  938  968 1010 1023 1040  1076  1096     0     0
#> [40,]   11    1  556  716  733  763  790  846  872   912   953  1018  1077
#> [41,]    6    1  705  908  910 1008 1068 1086 1096     0     0     0     0
#> [42,]    9    1  598  834  843  880  973  983 1009  1032  1066  1096     0
#> [43,]   10    1  415  581  628  842  845  860  901   959   985  1083  1096
#> [44,]    8    1  621  826  876  928  932  969  972  1064  1096     0     0
#> [45,]    8    1  668  904  919  993  995 1006 1058  1086  1096     0     0
#> [46,]   12    1  415  561  717  828  892  931  932   988   996  1030  1059
#> [47,]    5    1  618  621  931  960  972 1096    0     0     0     0     0
#> [48,]    6    1  596  642  694 1007 1056 1086 1096     0     0     0     0
#> [49,]   12    1  548  621  642  792  834  841  879   912   981   997  1084
#> [50,]   11    1  590  613  713  738  784  856  898   971   994  1006  1018
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]     0     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]  1096     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]  1046  1078  1096     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]  1053  1096     0     0     0     0     0
#> [22,]  1009  1040  1057  1060  1086  1096     0
#> [23,]  1044  1082  1096     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]  1077  1096     0     0     0     0     0
#> [26,]  1096     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]  1063  1096     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]  1096     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]  1082  1096     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]  1086  1096     0     0     0     0     0
#> [50,]  1096     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  761  780  809  837  858  859  914  929  935  959  964  987 1022
#> [16] 1036 1059 1064 1067 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   95  147  384  494  641  680  727  808  828  890  910  923  950
#> [16] 1013 1028 1038 1062 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  350  380  601  681  735  798  810  832  871  873  969 1023 1027
#> [16] 1040 1052 1069 1077 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1    5  247  603  627  676  736  793  802  850  910  919 1004 1016
#> [16] 1033 1034 1065 1084 1096
sim_1_cp_BMDL(exceedances(bogota_pm))
#>  [1]   17    1   24   36   54   61   62  167  173  284  317  370  378  400  417
#> [16]  507  592  775  909 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  488  557  688  810  813  875  963   968   983   989   991
#>  [2,]   17    1  722  728  798  841  844  861  868   900   919   923   946
#>  [3,]   17    1  271  488  621  780  791  842  848   851   860   887   900
#>  [4,]   17    1  553  566  575  589  722  748  831   834   853   892   905
#>  [5,]   17    1  609  647  654  663  678  688  798   817   826   901   905
#>  [6,]   17    1   20  581  729  836  838  879  887   891   919   959   975
#>  [7,]   17    1  755  822  828  860  870  884  916   966   982   987   995
#>  [8,]   16    1  557  605  628  642  763  784  878   881   901   941   961
#>  [9,]   17    1  354  553  659  672  791  794  841   848   860   901  1007
#> [10,]   17    1  233  680  798  825  831  842  849   856   886   897   940
#> [11,]   17    1  571  711  882  883  886  900  910   938   949   959   989
#> [12,]   17    1  719  733  844  851  871  879  890   943   984   986   996
#> [13,]   17    1  604  619  704  771  813  817  831   895   926   939   969
#> [14,]   17    1  614  642  672  706  785  868  891   940   943   955   968
#> [15,]   17    1  566  623  654  833  845  862  907   908   928   930   945
#> [16,]   17    1  553  628  642  681  697  753  844   864   872   877   894
#> [17,]   17    1  488  553  668  722  753  825  837   894   953   995  1014
#> [18,]   13    1  562  571  705  843  856  935  968   996  1021  1040  1052
#> [19,]   17    1  561  668  719  741  768  839  849   862   866   912   914
#> [20,]   15    1  614  657  706  713  784  798  859   876   885   994  1025
#> [21,]   17    1  791  798  803  817  828  834  844   854   861   888   891
#> [22,]   17    1  561  704  716  738  741  746  780   851   885   946   950
#> [23,]   17    1  226  271  562  581  749  845  865   866   895   898   920
#> [24,]   15    1  488  571  619  672  713  727  794   824   854   867   924
#> [25,]   12    1  271  753  763  832  882  885  915   979   995  1005  1041
#> [26,]   17    1  590  615  633  761  777  823  849   872   886   889   892
#> [27,]   17    1  488  634  655  698  716  746  826   831   884   893   900
#> [28,]   17    1  354  589  798  830  856  895  908   937   957   963   988
#> [29,]   17    1   20   60  590  614  647  713  763   840   841   849   911
#> [30,]   17    1  604  609  656  727  777  817  836   903   929   971   977
#> [31,]   17    1   51  556  589  633  716  746  761   858   862   898   904
#> [32,]   17    1  598  629  785  843  852  891  893   898   908   930   936
#> [33,]   17    1  654  694  728  748  841  857  870   890   893   948  1011
#> [34,]   17    1  553  621  656  822  831  842  885   890   904   921   928
#> [35,]   17    1  221  548  680  738  825  838  866   877   885   892   932
#> [36,]   17    1  604  656  687  768  803  830  831   872   883   949   959
#> [37,]   17    1  221  621  647  753  774  826  864   871   897   941   974
#> [38,]   17    1   51  566  590  615  618  646  663   711   794   890   930
#> [39,]   17    1  271  308  621  634  663  727  803   817   858   875   886
#> [40,]   17    1  581  705  849  854  872  893  918   926   940   943  1024
#> [41,]   17    1   60  604  623  647  719  722  780   792   836   853   859
#> [42,]   17    1  677  755  848  859  882  890  894   911   923   935   963
#> [43,]   17    1  583  642  677  687  771  798  823   868   876   910   920
#> [44,]   17    1  488  598  647  681  749  832  862   894   936   978   987
#> [45,]   17    1  571  680  705  741  803  822  846   866   868   888   915
#> [46,]   17    1   20  548  575  613  681  708  784   839   843   856   910
#> [47,]   17    1  609  681  688  765  803  842  843   870   919   928   930
#> [48,]   17    1  226  561  604  785  798  809  823   865   879   900   919
#> [49,]   17    1  628  668  671  708  763  785  810   822   825   854   868
#> [50,]   17    1  655  688  711  719  851  911  948   988  1000  1012  1013
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   995  1007  1012  1017  1032  1091  1096
#>  [2,]   966  1015  1024  1052  1060  1092  1096
#>  [3,]   950   973  1043  1052  1056  1094  1096
#>  [4,]   949   950  1006  1014  1055  1080  1096
#>  [5,]   911   946   951  1055  1058  1069  1096
#>  [6,]   977  1009  1043  1049  1053  1082  1096
#>  [7,]  1003  1044  1070  1071  1093  1095  1096
#>  [8,]   977  1020  1029  1059  1080  1096     0
#>  [9,]  1025  1044  1063  1072  1083  1090  1096
#> [10,]   961  1011  1018  1029  1035  1063  1096
#> [11,]  1001  1068  1070  1073  1086  1094  1096
#> [12,]  1041  1053  1054  1072  1090  1094  1096
#> [13,]   984  1000  1012  1025  1052  1069  1096
#> [14,]   978   983  1003  1057  1065  1077  1096
#> [15,]   980   987   999  1016  1020  1085  1096
#> [16,]   905   985  1020  1048  1082  1085  1096
#> [17,]  1021  1053  1071  1083  1084  1089  1096
#> [18,]  1076  1090  1096     0     0     0     0
#> [19,]   944   970  1002  1056  1068  1083  1096
#> [20,]  1034  1041  1045  1049  1096     0     0
#> [21,]   964  1021  1032  1056  1067  1080  1096
#> [22,]   971   977  1021  1037  1081  1087  1096
#> [23,]   957   978  1004  1026  1041  1049  1096
#> [24,]   945   977  1018  1089  1096     0     0
#> [25,]  1094  1096     0     0     0     0     0
#> [26,]   905   926   957   972  1063  1079  1096
#> [27,]   903   936   979   987  1054  1083  1096
#> [28,]  1005  1010  1052  1067  1076  1082  1096
#> [29,]   918   934   990  1009  1031  1047  1096
#> [30,]   980   982  1006  1028  1064  1089  1096
#> [31,]   935   950   953  1033  1059  1066  1096
#> [32,]   944   970   979  1000  1015  1021  1096
#> [33,]  1019  1024  1036  1037  1044  1045  1096
#> [34,]   929   961   979  1012  1046  1075  1096
#> [35,]   957   960  1007  1021  1062  1088  1096
#> [36,]   966   974   988  1048  1082  1094  1096
#> [37,]   979   985  1019  1062  1073  1077  1096
#> [38,]   953   982   992  1005  1048  1083  1096
#> [39,]   891   947   971   986   998  1079  1096
#> [40,]  1026  1035  1036  1058  1060  1087  1096
#> [41,]   863   928   945   978  1087  1091  1096
#> [42,]  1007  1008  1025  1072  1078  1080  1096
#> [43,]   929   933   988  1047  1068  1084  1096
#> [44,]  1024  1025  1044  1067  1086  1090  1096
#> [45,]   933   949   967   988  1002  1063  1096
#> [46,]   948   975  1009  1011  1013  1016  1096
#> [47,]   939   951   988   992  1010  1059  1096
#> [48,]   949  1007  1008  1052  1070  1073  1096
#> [49,]   934  1006  1026  1033  1036  1076  1096
#> [50,]  1018  1030  1043  1044  1076  1077  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986