Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    8    1  774  792  857  868  926  995 1018  1049  1096     0     0
#>  [2,]    8    1  749  833  897  994 1023 1027 1035  1051  1096     0     0
#>  [3,]    5    1  647  657 1002 1015 1086 1096    0     0     0     0     0
#>  [4,]    7    1  583  596  884  908  962  984 1071  1096     0     0     0
#>  [5,]    6    1  704  836  896  918 1004 1066 1096     0     0     0     0
#>  [6,]   13    1  583  681  697  711  771  840  850   882   893   894   943
#>  [7,]    8    1  717  792  813  878  899  940 1048  1050  1096     0     0
#>  [8,]    9    1  415  705  875  877  891  898  921   979  1026  1096     0
#>  [9,]    6    1  575  687  713  921  949 1051 1096     0     0     0     0
#> [10,]   10    1  561  867  880  962  972 1002 1009  1015  1041  1068  1096
#> [11,]    8    1  888  907  911  929 1007 1008 1026  1064  1096     0     0
#> [12,]   10    1  755  897  929  945  967  982  989  1011  1038  1065  1096
#> [13,]    7    1  716  826  853  969  975  980 1046  1096     0     0     0
#> [14,]    8    1  613  716  883  924  968 1013 1031  1063  1096     0     0
#> [15,]   11    1  613  687  719  851  881  912  934   968   971   982  1031
#> [16,]    6    1  883  943  944  973 1031 1062 1096     0     0     0     0
#> [17,]   11    1  613  841  852  904  945  960  975  1015  1025  1080  1086
#> [18,]   11    1  605  761  830  833  931  981 1007  1046  1048  1083  1086
#> [19,]    7    1  846  888  950  973  985 1012 1060  1096     0     0     0
#> [20,]    5    1  924  926  966 1002 1079 1096    0     0     0     0     0
#> [21,]   16    1  583  629  694  771  774  830  833   851   872   892   941
#> [22,]    8    1  415  705  790  914  918  931 1021  1054  1096     0     0
#> [23,]    5    1  677  713  935  941 1074 1096    0     0     0     0     0
#> [24,]    6    1  556  822  901  919  996 1029 1096     0     0     0     0
#> [25,]    8    1  618  649  817  872  888  987 1046  1078  1096     0     0
#> [26,]    8    1  654  687  719  791  837  919 1021  1086  1096     0     0
#> [27,]    9    1  722  741  856  959 1006 1009 1025  1064  1078  1096     0
#> [28,]    8    1  556  557  746  938  983 1037 1056  1084  1096     0     0
#> [29,]    9    1  557  589  748  763  828  839  972  1083  1086  1096     0
#> [30,]   10    1  590  842  908  911  944  974 1002  1020  1037  1086  1096
#> [31,]    8    1  688  830  833  836  970 1018 1022  1058  1096     0     0
#> [32,]    7    1  415  706  761  910  991 1067 1083  1096     0     0     0
#> [33,]   13    1  657  727  817  839  954  983  996  1016  1023  1039  1053
#> [34,]   11    1  680  722  784  825  879  886  960   974   991  1024  1085
#> [35,]    9    1  655  746  768  920  929  934  938   989  1086  1096     0
#> [36,]    9    1  598  680  809  848  849  900  985  1002  1058  1096     0
#> [37,]    8    1  655  774  930  979  992 1014 1046  1086  1096     0     0
#> [38,]    7    1  621  649  830  919  991 1044 1048  1096     0     0     0
#> [39,]   10    1  415  589  694  708  774  809  859   879   984  1086  1096
#> [40,]    7    1  415  619  858  974  976 1016 1025  1096     0     0     0
#> [41,]    8    1  583  711  738  834  971 1008 1069  1086  1096     0     0
#> [42,]   12    1  609  618  725  832  934  947  981   987  1027  1028  1031
#> [43,]    7    1  904  993  996 1008 1067 1072 1086  1096     0     0     0
#> [44,]    8    1  615  645  657  687  831  852 1078  1086  1096     0     0
#> [45,]    9    1  629  878  892  948  961  992 1021  1026  1086  1096     0
#> [46,]   11    1  633  725  761  853  897  924 1001  1016  1046  1058  1067
#> [47,]    9    1  619  655  837  893  943  949  987   994  1061  1096     0
#> [48,]   12    1  561  562  657  748  863  868  879   994  1003  1005  1010
#> [49,]   11    1  615  646  659  672  749  765  880   929   955  1008  1041
#> [50,]    9    1  719  791  916  924  994 1007 1009  1078  1086  1096     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]     0     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]   963  1022  1096     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]  1096     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]  1096     0     0     0     0     0     0
#> [18,]  1096     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]   948   962  1007  1033  1069  1096     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]     0     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]  1062  1086  1096     0     0     0     0
#> [34,]  1096     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]  1050  1096     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]  1096     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]  1016  1096     0     0     0     0     0
#> [49,]  1096     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  614  623  642  672  706  785  891  940  943  955  968  978  983
#> [16] 1003 1057 1065 1077 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   62  201  524  591  741  743  799  802  826  863  884  903  910
#> [16]  990 1006 1021 1085 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  382  451  457  493  729  738  741  761  773  849  960  964  981
#> [16]  999 1039 1084 1085 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   15    1  216  687  728  758  821  831  838  991 1001 1017 1044 1049 1078
#> [16] 1079 1080 1096    0    0
sim_1_cp_BMDL(exceedances(bogota_pm))
#>  [1]   17    1  120  213  308  395  443  462  471  506  601  666  681  753  757
#> [16]  782  801  977 1055 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  561  571  668  768  839  849  856   866   935   944   968
#>  [2,]   17    1  614  657  706  719  741  784  859   862   876   885   912
#>  [3,]   17    1  713  791  798  803  817  828  834   888   991   994  1021
#>  [4,]   17    1  561  704  738  741  780  844  854   861   885   891   950
#>  [5,]   17    1  226  271  562  581  716  746  851   866   895   898   946
#>  [6,]   17    1  488  619  672  713  727  749  794   824   845   854   867
#>  [7,]   17    1  271  748  753  763  832  849  882   885   886   915   926
#>  [8,]   17    1  488  590  633  655  761  777  823   872   884   889   892
#>  [9,]   16    1  634  698  716  746  826  856  893   895   900   936   937
#> [10,]   17    1  354  589  713  763  798  830  841   908   911   957   963
#> [11,]   14    1   60  590  605  614  647  840  849   903   918   934   990
#> [12,]   17    1  604  609  656  727  777  817  836   929   935   971   980
#> [13,]   15    1   51  556  589  633  716  746  761   858   862   898   904
#> [14,]   17    1  598  629  785  843  852  891  893   898   908   930   936
#> [15,]   17    1  654  694  728  748  841  857  870   890   893   948  1011
#> [16,]   17    1  553  621  656  822  831  842  885   890   904   921   928
#> [17,]   17    1  221  548  680  738  825  838  866   877   885   892   932
#> [18,]   17    1  604  656  687  768  803  830  831   872   883   949   959
#> [19,]   17    1  221  621  647  753  774  826  864   871   897   941   974
#> [20,]   17    1   51  566  590  615  618  646  663   711   794   890   930
#> [21,]   17    1  271  308  621  634  663  727  803   817   858   875   886
#> [22,]   17    1  581  705  849  854  872  893  918   926   940   943  1024
#> [23,]   17    1   60  604  623  647  719  722  780   792   836   853   859
#> [24,]   17    1  677  755  848  859  882  890  894   911   923   935   963
#> [25,]   17    1  583  642  677  687  771  798  823   868   876   910   920
#> [26,]   17    1  488  598  647  681  749  832  862   894   936   978   987
#> [27,]   17    1  571  680  705  741  803  822  846   866   868   888   915
#> [28,]   17    1   20  548  575  613  681  708  784   839   843   856   910
#> [29,]   17    1  609  681  688  765  803  842  843   870   919   928   930
#> [30,]   17    1  226  561  604  785  798  809  823   865   879   900   919
#> [31,]   17    1  628  668  671  708  763  785  810   822   825   854   868
#> [32,]   17    1  655  688  711  719  851  911  948   988  1000  1012  1013
#> [33,]   17    1  583  678  688  697  844  938  943   950   984   988  1002
#> [34,]   13    1  615  694  729  741  753  765  863   878   889   932   954
#> [35,]   17    1  628  645  657  697  765  817  849   882   912   934   944
#> [36,]   17    1   60  233  561  628  717  882  887   891   923   949   964
#> [37,]   17    1   20  354  746  859  872  875  911   953   959   960   962
#> [38,]   17    1   51  226  548  609  677  838  870   933   939   999  1000
#> [39,]   17    1  548  561  668  688  836  885  907   920   936   970   972
#> [40,]   17    1  604  605  666  668  704  749  785   791   825   830  1009
#> [41,]   17    1  668  713  716  748  810  850  889   900   935   991   994
#> [42,]   17    1  271  548  589  746  780  867  883   886   918   997  1009
#> [43,]   17    1   51   60  575  590  604  708  741   834   851   866   870
#> [44,]   17    1  598  621  633  645  719  725  771   780   798   822   842
#> [45,]   15    1  633  706  840  846  877  907  912   915   937   963   983
#> [46,]   17    1  415  668  688  694  706  853  878   880   894   963   989
#> [47,]   17    1  556  589  646  663  681  777  846   870   901   912   940
#> [48,]   17    1  226  596  645  647  680  784  860   877   889   899   904
#> [49,]   17    1  233  623  633  809  813  854  877   926   929   937   940
#> [50,]   17    1  609  618  656  681  719  828  877   945   949   961   977
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   999  1002  1056  1076  1083  1089  1096
#>  [2,]   914  1025  1034  1049  1068  1076  1096
#>  [3,]  1032  1041  1045  1056  1067  1094  1096
#>  [4,]   964   971  1021  1080  1081  1087  1096
#>  [5,]   957   977   978  1004  1026  1049  1096
#>  [6,]   920   924   945   977  1018  1041  1096
#>  [7,]   979   995  1005  1041  1067  1093  1096
#>  [8,]   903   905   957   972  1063  1079  1096
#>  [9,]   979   987  1028  1054  1076  1096     0
#> [10,]   988  1010  1052  1067  1071  1087  1096
#> [11,]  1009  1031  1047  1096     0     0     0
#> [12,]   982  1006  1028  1059  1064  1091  1096
#> [13,]   950  1033  1059  1066  1096     0     0
#> [14,]   944   970   979  1000  1015  1021  1096
#> [15,]  1019  1024  1036  1037  1044  1045  1096
#> [16,]   929   961   979  1012  1046  1075  1096
#> [17,]   957   960  1007  1021  1062  1088  1096
#> [18,]   966   974   988  1048  1082  1094  1096
#> [19,]   979   985  1019  1062  1073  1077  1096
#> [20,]   953   982   992  1005  1048  1083  1096
#> [21,]   891   947   971   986   998  1079  1096
#> [22,]  1026  1035  1036  1058  1060  1087  1096
#> [23,]   863   928   945   978  1087  1091  1096
#> [24,]  1007  1008  1025  1072  1078  1080  1096
#> [25,]   929   933   988  1047  1068  1084  1096
#> [26,]  1024  1025  1044  1067  1086  1090  1096
#> [27,]   933   949   967   988  1002  1063  1096
#> [28,]   948   975  1009  1011  1013  1016  1096
#> [29,]   939   951   988   992  1010  1059  1096
#> [30,]   949  1007  1008  1052  1070  1073  1096
#> [31,]   934  1006  1026  1033  1036  1076  1096
#> [32,]  1018  1030  1043  1044  1076  1077  1096
#> [33,]  1015  1021  1033  1044  1065  1090  1096
#> [34,]  1054  1073  1096     0     0     0     0
#> [35,]   962   972  1030  1041  1068  1078  1096
#> [36,]  1002  1018  1020  1076  1080  1093  1096
#> [37,]   984   999  1016  1017  1029  1060  1096
#> [38,]  1015  1016  1024  1045  1058  1078  1096
#> [39,]   985  1035  1043  1044  1083  1088  1096
#> [40,]  1021  1028  1044  1048  1062  1064  1096
#> [41,]  1045  1052  1060  1064  1070  1086  1096
#> [42,]  1014  1023  1070  1081  1087  1092  1096
#> [43,]   950   971  1016  1029  1037  1066  1096
#> [44,]   882   897   992  1008  1043  1087  1096
#> [45,]  1005  1068  1072  1075  1096     0     0
#> [46,]   992  1009  1019  1064  1072  1088  1096
#> [47,]   957   960   982  1019  1047  1071  1096
#> [48,]   960   987   988  1009  1024  1092  1096
#> [49,]   982  1023  1041  1054  1072  1085  1096
#> [50,]   980  1026  1047  1056  1070  1086  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986