Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   11    1  716  733  749  785  813  896  928   953   986  1013  1055
#>  [2,]   10    1  605  647  791  858  871  919  947   948  1037  1053  1096
#>  [3,]    9    1  488  589  642  755  809  824  850   857   999  1096     0
#>  [4,]   12    1  615  634  645  711  817  853  864   875   887   934   973
#>  [5,]    5    1  634  678 1020 1036 1056 1096    0     0     0     0     0
#>  [6,]    9    1  628  765  872  967  977 1034 1047  1055  1058  1096     0
#>  [7,]    8    1  415  642  713  853  871  973  995  1050  1096     0     0
#>  [8,]    8    1  649  697  761  933  962  969  977   985  1096     0     0
#>  [9,]    6    1  634  903  943  974 1047 1067 1096     0     0     0     0
#> [10,]   11    1  619  792  851  915  919  946  949   963  1002  1048  1086
#> [11,]   10    1  716  828  857  896  920 1002 1010  1025  1042  1067  1096
#> [12,]   10    1  548  605  765  878  892  899  947   981   989  1078  1096
#> [13,]    8    1  633  746  950  969  984  992 1033  1038  1096     0     0
#> [14,]   10    1  415  596  830  842  851  887  921   981  1047  1086  1096
#> [15,]    8    1  556  575  649  872  903  943 1046  1066  1096     0     0
#> [16,]    8    1  609  672  705  722  907 1011 1029  1086  1096     0     0
#> [17,]    7    1  589  706  716  784  794  997 1030  1096     0     0     0
#> [18,]    9    1  741  803  830  878  900  914  954   966   979  1096     0
#> [19,]    7    1  575  791  905  923  933  949 1048  1096     0     0     0
#> [20,]    9    1  619  628  687  716  984  986 1031  1045  1081  1096     0
#> [21,]    8    1  614  704  817  826  843  933  947   971  1096     0     0
#> [22,]   13    1  659  780  838  849  877  879  881   928   946   949   967
#> [23,]    8    1  557  649  943  983 1012 1035 1047  1053  1096     0     0
#> [24,]   12    1  621  657  687  727  791  831  832   997  1004  1056  1072
#> [25,]    9    1  706  716  746  859  881  948 1012  1059  1086  1096     0
#> [26,]    8    1  647  688  877  892  903  916 1034  1045  1096     0     0
#> [27,]    9    1  688  722  780  822  879  985 1054  1064  1078  1096     0
#> [28,]   10    1  621  794  841  914  969  977  993  1010  1034  1066  1096
#> [29,]    8    1  488  583  755  784  872  946 1037  1060  1096     0     0
#> [30,]   12    1  556  633  681  842  850  941  959   969   970  1008  1013
#> [31,]    9    1  604  634  755  761  846  924 1018  1046  1057  1096     0
#> [32,]    9    1  845  881  882  891  912 1003 1059  1085  1086  1096     0
#> [33,]    9    1  553  561  623  645  755 1009 1023  1045  1074  1096     0
#> [34,]   10    1  598  657  681  717  810  843  929   947  1030  1053  1096
#> [35,]    8    1  668  803  856  863  886  914  919   984  1096     0     0
#> [36,]    8    1  615  722  794  850  860  966  979  1086  1096     0     0
#> [37,]    8    1  654  748  862  900  939  966 1047  1057  1096     0     0
#> [38,]    5    1  753  825  943  951 1013 1096    0     0     0     0     0
#> [39,]    7    1  596  711  809  933  972  976 1002  1096     0     0     0
#> [40,]    7    1  777  899  951  953  992 1046 1078  1096     0     0     0
#> [41,]    6    1  415  621  634  755  803  828 1096     0     0     0     0
#> [42,]    7    1  415  571  719  844  912  970  993  1096     0     0     0
#> [43,]   12    1  590  713  761  791  810  831  898   962   968   980  1035
#> [44,]   13    1  415  553  562  717  790  826  836   857   859   883   947
#> [45,]    7    1  415  642  681  688  853  914 1024  1096     0     0     0
#> [46,]    9    1  562  694  838  882  926  932  994  1040  1086  1096     0
#> [47,]    8    1  645  725  834  840  844  894  931  1003  1096     0     0
#> [48,]    7    1  598  668  694  887  933  976 1083  1096     0     0     0
#> [49,]   10    1  557  609  642  704  753  831  843   900   949  1016  1096
#> [50,]    7    1  415  634  798  840  947 1065 1074  1096     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  1096     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]  1016  1096     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]  1096     0     0     0     0     0     0
#> [11,]     0     0     0     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]  1039  1064  1096     0     0     0     0
#> [23,]     0     0     0     0     0     0     0
#> [24,]  1086  1096     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]  1051  1096     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]     0     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]  1086  1096     0     0     0     0     0
#> [44,]   953  1045  1096     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  725  809  840  857  864  876  889  937  961  964  970  995 1008
#> [16] 1014 1033 1039 1070 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   48  157  275  336  390  393  469  572  664  828  859  930  937
#> [16]  954  961 1039 1068 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  304  374  467  484  531  532  764  766  784  817  831  835  868
#> [16]  965 1006 1043 1053 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1  380  419  626  629  674  678  728  763  842  857  862  866  896
#> [16]  952 1019 1048 1054 1096
sim_1_cp_BMDL(exceedances(bogota_pm))
#>  [1]   17    1   64   93  120  138  258  426  429  456  464  496  536  605  640
#> [16]  644 1005 1042 1083 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  557  590  613  615  713  771  809   836   895   898   905
#>  [2,]   17    1  581  605  654  678  706  725  862   879   928   935   949
#>  [3,]   17    1  649  656  725  823  826  859  903   919   928   953   987
#>  [4,]   17    1  581  583  614  628  629  663  666   824   862   926   950
#>  [5,]   17    1  590  609  623  768  791  798  857   886   897   912   915
#>  [6,]   16    1  488  704  713  753  834  921  945   961   972   984   993
#>  [7,]   17    1  645  654  717  748  817  849  857   862   865   887   898
#>  [8,]   17    1  677  729  768  784  831  834  836   843   858   860   862
#>  [9,]   17    1  553  677  694  741  780  817  833   837   839   849   877
#> [10,]   17    1  642  680  755  803  813  831  852   860   910   978   980
#> [11,]   17    1  557  629  704  792  843  844  882   901   911   916   940
#> [12,]   17    1   20  415  562  687  722  755  777   831   872   929   990
#> [13,]   17    1  562  566  646  717  733  850  861   862   870   883   948
#> [14,]   15    1   60  688  706  727  748  840  843   885   959   973   975
#> [15,]   17    1  354  553  566  613  655  708  727   842   919   979   980
#> [16,]   17    1  553  561  655  733  774  784  791   817   857   875   916
#> [17,]   17    1   51  556  566  621  623  659  666   810   837   844   848
#> [18,]   17    1  571  575  609  613  828  842  862   868   875   900   904
#> [19,]   17    1  621  642  697  780  825  838  839   846   854   897   968
#> [20,]   17    1   51  566  711  765  852  878  882   910   933   941   945
#> [21,]   17    1  583  649  687  708  763  780  794   865   866   893   934
#> [22,]   17    1  415  605  642  729  755  761  831   907   932   944   988
#> [23,]   17    1  233  581  615  659  687  749  825   852   860   908   928
#> [24,]   17    1  680  694  765  803  825  830  831   842   862   990  1004
#> [25,]   17    1   60  596  654  672  694  768  933   944   949   976   995
#> [26,]   17    1  548  596  729  765  849  868  872   888   892   910   915
#> [27,]   17    1  575  618  629  654  704  725  809   867   886   923   968
#> [28,]   17    1  354  566  596  663  677  753  768   903   918   929   935
#> [29,]   17    1  354  649  733  780  935  946  947   948   996  1002  1010
#> [30,]   17    1  553  557  566  642  774  784  852   868   960   971   975
#> [31,]   17    1  221  589  659  866  892  904  912   931   934   954   973
#> [32,]   17    1  308  488  708  717  794  813  837   893   895   937   938
#> [33,]   17    1  561  642  790  824  849  851  852   880   907   928  1019
#> [34,]   17    1  621  705  826  851  876  918  926   930   932   940   959
#> [35,]   16    1  571  654  672  716  728  753  771   774   792   966   976
#> [36,]   17    1  621  706  746  748  771  863  875   890   896   908   912
#> [37,]   17    1  415  680  705  725  792  831  841   867   904   923   938
#> [38,]   15    1  354  649  688  784  813  845  866   881   940   979  1008
#> [39,]   17    1   20  226  548  613  615  672  698   840   842   872   935
#> [40,]   14    1   51  548  583  657  834  850  918   931   938   951   962
#> [41,]   17    1  354  589  645  694  755  774  834   867   919   957   984
#> [42,]   17    1  233  561  598  647  722  746  823   825   828   850   926
#> [43,]   17    1  566  618  642  833  875  920  933   954   996  1001  1004
#> [44,]   17    1  557  698  774  845  884  898  904   933   961   968   982
#> [45,]   17    1  663  677  761  768  837  842  843   887   888   911   948
#> [46,]   17    1  589  716  741  866  918  934  946   963   973   986  1008
#> [47,]   17    1  488  654  680  711  733  790  809   837   894   898   949
#> [48,]   17    1  575  729  817  825  848  868  878   887   901   935   936
#> [49,]   17    1  571  634  705  711  845  881  897   932   934   970   972
#> [50,]   17    1  415  557  646  659  680  790  810   862   865   871   882
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   908   918   930   968  1011  1024  1096
#>  [2,]   979  1007  1031  1038  1065  1084  1096
#>  [3,]  1002  1016  1033  1041  1066  1072  1096
#>  [4,]   974  1022  1040  1049  1068  1090  1096
#>  [5,]   943   948   964  1002  1040  1078  1096
#>  [6,]  1011  1046  1073  1078  1084  1096     0
#>  [7,]   928   929   940   967   983  1083  1096
#>  [8,]   863  1002  1015  1027  1033  1060  1096
#>  [9,]   889   924   982  1008  1050  1092  1096
#> [10,]  1006  1009  1013  1070  1073  1076  1096
#> [11,]   978   986  1044  1045  1050  1078  1096
#> [12,]  1012  1018  1032  1043  1071  1076  1096
#> [13,]   986   996  1021  1022  1033  1095  1096
#> [14,]   997  1025  1059  1068  1096     0     0
#> [15,]  1001  1012  1044  1060  1069  1079  1096
#> [16,]   936   939   984  1002  1027  1053  1096
#> [17,]   856   877   892   893  1047  1064  1096
#> [18,]   981   987   996  1048  1059  1070  1096
#> [19,]  1008  1012  1024  1044  1073  1085  1096
#> [20,]   951   978  1001  1023  1055  1080  1096
#> [21,]   972   973   999  1022  1027  1033  1096
#> [22,]  1005  1020  1030  1036  1069  1090  1096
#> [23,]   936  1009  1015  1053  1080  1088  1096
#> [24,]  1006  1057  1068  1071  1075  1080  1096
#> [25,]  1033  1063  1064  1065  1066  1074  1096
#> [26,]   984   986   993  1007  1056  1094  1096
#> [27,]   999  1003  1006  1044  1076  1087  1096
#> [28,]   964   995  1014  1071  1077  1085  1096
#> [29,]  1023  1024  1026  1029  1069  1090  1096
#> [30,]   988   994  1003  1005  1030  1077  1096
#> [31,]  1025  1029  1066  1074  1086  1092  1096
#> [32,]   959   966   980  1005  1008  1092  1096
#> [33,]  1045  1050  1067  1073  1088  1089  1096
#> [34,]  1006  1036  1045  1061  1084  1089  1096
#> [35,]   989   999  1026  1052  1065  1096     0
#> [36,]   935   986  1015  1076  1089  1094  1096
#> [37,]   940   947   951   953   959  1037  1096
#> [38,]  1042  1043  1052  1060  1096     0     0
#> [39,]  1013  1034  1042  1046  1052  1067  1096
#> [40,]   979  1014  1095  1096     0     0     0
#> [41,]  1004  1015  1021  1075  1076  1087  1096
#> [42,]   984   995   997  1015  1085  1087  1096
#> [43,]  1020  1023  1031  1051  1060  1064  1096
#> [44,]  1011  1050  1053  1070  1072  1080  1096
#> [45,]   951   971   985   993  1041  1095  1096
#> [46,]  1014  1015  1017  1023  1063  1070  1096
#> [47,]   955   980  1000  1004  1049  1067  1096
#> [48,]   950   970   984  1017  1020  1039  1096
#> [49,]   976   989   992  1052  1061  1076  1096
#> [50,]   889   912   941   980  1044  1072  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986