Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_m+1; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]    8    1  828  870  885  893  955  975  994  1080  1096     0     0
#>  [2,]    9    1  672  768  840  880  923  981 1031  1043  1071  1096     0
#>  [3,]    9    1  609  765  884  936  982 1024 1063  1077  1079  1096     0
#>  [4,]   11    1  656  798  865  896  915  928 1002  1029  1043  1063  1070
#>  [5,]   12    1  415  628  803  951  954  967  976   995  1000  1040  1076
#>  [6,]    7    1  548  697  824  872  959  964  992  1096     0     0     0
#>  [7,]   11    1  719  755  810  842  903  934  990  1018  1040  1055  1086
#>  [8,]   10    1  780  813  824  860  877  903  919  1019  1079  1086  1096
#>  [9,]   11    1  654  681  733  831  867  871  944   973   982  1047  1081
#> [10,]    7    1  623  711  852  915  984 1052 1065  1096     0     0     0
#> [11,]   13    1  557  598  677  687  785  880  943   966   971   990  1007
#> [12,]    5    1  596  893  905  948  979 1096    0     0     0     0     0
#> [13,]    9    1  748  834  841  861  875  892  896   898   940  1096     0
#> [14,]   17    1  557  598  678  719  729  741  777   794   809   810   837
#> [15,]   10    1  575  761  791  832  838  975 1006  1018  1030  1086  1096
#> [16,]    3    1  415  973 1008 1096    0    0    0     0     0     0     0
#> [17,]    8    1  748  803  866  920  973 1002 1049  1050  1096     0     0
#> [18,]   10    1  589  647  831  832  868  870  984   985  1065  1084  1096
#> [19,]    9    1  654  822  900  914  938  992  996  1045  1086  1096     0
#> [20,]    8    1  488  623  687  872  898  947 1030  1048  1096     0     0
#> [21,]   10    1  566  590  794  830  832  868  962   982  1003  1086  1096
#> [22,]   10    1  571  623  681  841  854  941  972  1001  1015  1040  1096
#> [23,]   11    1  619  646  704  831  863  866  893   945   995  1019  1082
#> [24,]   10    1  716  839  857  872  875  879  891   934  1008  1035  1096
#> [25,]    6    1  843  850  895  983 1001 1014 1096     0     0     0     0
#> [26,]    9    1  717  763  833  853  916  946  973  1066  1077  1096     0
#> [27,]   11    1  621  725  746  839  938  959  991  1000  1002  1014  1030
#> [28,]    9    1  557  590  794  817  871  972  978  1005  1086  1096     0
#> [29,]   10    1  415  557  681  828  854  863  912   982  1025  1044  1096
#> [30,]   13    1  557  605  649  657  705  748  765   836   936   959  1000
#> [31,]    6    1  628  886  941 1017 1028 1086 1096     0     0     0     0
#> [32,]   12    1  415  798  828  851  883  908  911   996  1007  1033  1053
#> [33,]    9    1  415  618  657  722  746  822  837   999  1002  1096     0
#> [34,]    6    1  415  831  976 1025 1038 1060 1096     0     0     0     0
#> [35,]    3    1  833  848  894 1096    0    0    0     0     0     0     0
#> [36,]    9    1  561  647  663  885  953  971 1011  1038  1086  1096     0
#> [37,]    6    1  557  749  833  929  943 1075 1096     0     0     0     0
#> [38,]   12    1  681  698  831  839  868  898  907  1000  1029  1063  1078
#> [39,]   11    1  605  697  784  843  867  893  903   923   941  1015  1039
#> [40,]    8    1  566  725  810  929  936  937  946   959  1096     0     0
#> [41,]    6    1  717  826  839  887  972 1034 1096     0     0     0     0
#> [42,]    9    1  677  727  825  908  918 1037 1059  1061  1077  1096     0
#> [43,]   11    1  784  790  832  940  959  963  972  1028  1029  1039  1075
#> [44,]    7    1  825  839  872  876 1002 1026 1039  1096     0     0     0
#> [45,]   11    1  761  839  842  843  895  943  955   963   982  1016  1080
#> [46,]    9    1  654  704  705  852  904 1019 1040  1060  1064  1096     0
#> [47,]    9    1  798  849  859  907  928  949 1007  1033  1039  1096     0
#> [48,]    9    1  645  746  768  853  929  947 1006  1018  1086  1096     0
#> [49,]   14    1  557  604  649  680  706  711  848   853   889   894   946
#> [50,]   12    1  706  727  792  822  852  888  908   997  1010  1014  1060
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]     0     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]  1096     0     0     0     0     0     0
#>  [5,]  1086  1096     0     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]  1096     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]  1096     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]  1043  1086  1096     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]   898   901   916   938   969  1080  1096
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]  1096     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]  1096     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]  1003  1071  1096     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]  1059  1096     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]  1086  1096     0     0     0     0     0
#> [39,]  1096     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]  1096     0     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]  1096     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]   966  1030  1063  1096     0     0     0
#> [50,]  1075  1096     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1  233  354  553  634  698  862  865  871  886  890  919  940  970
#> [16]  971  988  992 1020 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   16   56  168  263  300  344  407  493  560  593  651  723  781
#> [16]  892  941  955 1005 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  382  408  464  483  525  529  605  744  786  793  799  873  917
#> [16]  987  991 1004 1092 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1  554  686  750  808  832  839  846  866  890  923  943  948  963
#> [16]  975  986 1018 1060 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  556  615  642  672  716  728  761   823   824   830   848
#>  [2,]   17    1  659  722  853  870  905  910  921   943   954   969   981
#>  [3,]   17    1  575  605  656  678  706  864  881   892   930   940   946
#>  [4,]   17    1  609  680  741  763  825  833  951   954   980   989  1000
#>  [5,]   17    1  618  619  642  705  716  719  755   785   843   887   901
#>  [6,]   17    1   20   51  609  725  728  790  836   894   900   924   945
#>  [7,]   17    1  557  566  848  853  861  880  893   918   962   975   986
#>  [8,]   17    1  556  583  589  613  694  746  834   864   865   918   953
#>  [9,]   16    1  415  671  774  842  889  908  914   943   954   979   981
#> [10,]   17    1  233  657  777  824  868  877  894   899   907   910   950
#> [11,]   17    1   60  671  677  763  837  853  878   928   932   946   992
#> [12,]   17    1  609  798  851  860  888  892  893   932   961   979   993
#> [13,]   14    1  596  654  727  768  785  861  870   877   894   963   970
#> [14,]   17    1  415  590  623  680  713  761  785   859   898   901   916
#> [15,]   17    1   60  657  694  698  830  858  875   876   903   924   931
#> [16,]   17    1  562  604  646  694  784  798  822   837   842   875   959
#> [17,]   17    1  271  642  645  678  708  780  839   851   868   870   892
#> [18,]   17    1  354  791  794  813  824  833  848   863   930   943   945
#> [19,]   15    1  614  628  656  704  711  822  833   836   868   885   896
#> [20,]   17    1  354  556  777  823  836  839  852   863   881   885   968
#> [21,]   17    1  226  583  623  642  728  733  810   842   879   896   920
#> [22,]   17    1  634  645  666  678  706  774  845   864   876   878   923
#> [23,]   17    1  561  618  671  687  719  852  860   935   936   954   970
#> [24,]   17    1  221  654  657  780  831  879  912   914   921   929   990
#> [25,]   17    1  561  614  645  657  688  813  850   860   977  1000  1002
#> [26,]   17    1  604  704  803  834  838  843  859   865   880   882   926
#> [27,]   17    1  415  623  629  677  840  863  891   899   936   946   963
#> [28,]   16    1  581  596  663  716  727  777  785   859   911   919   937
#> [29,]   17    1  590  613  657  774  803  813  839   850   889   934   938
#> [30,]   17    1  221  354  561  687  697  704  785   862   866   924   979
#> [31,]   17    1  803  872  878  889  905  915  985   996  1000  1018  1043
#> [32,]   17    1  604  609  681  698  719  768  780   784   798   813   862
#> [33,]   17    1  581  605  618  623  656  729  813   844   870   878   955
#> [34,]   17    1  556  605  678  680  794  817  852   878   899   977  1028
#> [35,]   17    1  604  646  678  680  761  846  851   867   882   899   935
#> [36,]   17    1  557  589  609  614  649  704  713   768   828   905   969
#> [37,]   17    1  583  598  621  698  708  838  858   860   894   916   970
#> [38,]   17    1   20  308  571  633  654  833  849   911   929   941   945
#> [39,]   17    1  647  705  803  833  837  859  871   916   923   933   945
#> [40,]   17    1  621  628  659  755  813  860  878   884   936   960   976
#> [41,]   17    1  589  655  666  668  705  706  832   834   843   903   954
#> [42,]   17    1  581  618  663  755  825  852  872   904   912   955   978
#> [43,]   17    1   20  666  755  761  777  831  851   889   933   936   979
#> [44,]   17    1   20  556  583  590  642  746  824   833   865   870   912
#> [45,]   17    1  656  755  825  839  851  876  924   928   943   978   986
#> [46,]   17    1  221  271  613  619  647  733  830   844   951   954   960
#> [47,]   17    1  415  615  716  832  842  872  876   904   912   939   948
#> [48,]   15    1  271  583  666  708  728  824  875   877   890   919   939
#> [49,]   17    1  604  728  813  848  856  877  928   944   946   990  1003
#> [50,]   17    1  556  557  654  671  681  688  727   728   834   908   940
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   923  1023  1041  1045  1077  1081  1096
#>  [2,]   989   991   993  1011  1070  1090  1096
#>  [3,]   959   996  1015  1034  1046  1085  1096
#>  [4,]  1004  1036  1040  1061  1074  1075  1096
#>  [5,]   962  1016  1021  1066  1073  1089  1096
#>  [6,]   969   971   988  1032  1037  1052  1096
#>  [7,]   988   991  1002  1020  1040  1052  1096
#>  [8,]  1035  1060  1072  1073  1076  1087  1096
#>  [9,]   995  1012  1041  1075  1077  1096     0
#> [10,]   955   979   996  1018  1048  1084  1096
#> [11,]   998  1015  1019  1039  1050  1081  1096
#> [12,]   996  1033  1041  1058  1060  1083  1096
#> [13,]   977   992  1060  1096     0     0     0
#> [14,]   921   983   986   996  1049  1068  1096
#> [15,]   946   962   989   996  1001  1075  1096
#> [16,]   975  1010  1015  1016  1046  1087  1096
#> [17,]   932   963   966   992  1003  1035  1096
#> [18,]   978   979   997  1003  1058  1093  1096
#> [19,]   959   982  1034  1051  1096     0     0
#> [20,]   988   998  1020  1038  1064  1077  1096
#> [21,]   932  1007  1013  1019  1061  1067  1096
#> [22,]   937   979  1029  1041  1050  1093  1096
#> [23,]   973   991  1002  1006  1034  1078  1096
#> [24,]  1015  1026  1047  1049  1087  1095  1096
#> [25,]  1003  1025  1026  1046  1050  1065  1096
#> [26,]   934   954   961  1002  1027  1035  1096
#> [27,]   964   985  1031  1044  1066  1078  1096
#> [28,]   964  1016  1021  1046  1069  1096     0
#> [29,]   950   981   987  1004  1031  1084  1096
#> [30,]  1009  1010  1020  1029  1033  1043  1096
#> [31,]  1059  1073  1077  1081  1091  1095  1096
#> [32,]   866   898   901  1026  1028  1038  1096
#> [33,]   990  1004  1023  1064  1073  1089  1096
#> [34,]  1030  1036  1040  1051  1083  1085  1096
#> [35,]   984  1003  1005  1036  1042  1057  1096
#> [36,]   970   990  1013  1029  1077  1080  1096
#> [37,]   978   984  1002  1004  1017  1020  1096
#> [38,]   994  1006  1026  1042  1048  1049  1096
#> [39,]   948   991   993  1031  1064  1068  1096
#> [40,]   984  1000  1024  1061  1064  1094  1096
#> [41,]   966   969  1012  1042  1058  1060  1096
#> [42,]  1004  1011  1027  1042  1076  1079  1096
#> [43,]   989  1001  1026  1027  1050  1060  1096
#> [44,]   955   972  1032  1040  1079  1088  1096
#> [45,]   991  1001  1007  1023  1024  1049  1096
#> [46,]   984   999  1041  1052  1060  1089  1096
#> [47,]   953   976   996  1052  1058  1078  1096
#> [48,]   963  1026  1075  1092  1096     0     0
#> [49,]  1030  1033  1034  1035  1051  1071  1096
#> [50,]  1009  1026  1040  1058  1064  1093  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986