Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   12    1  629  672  716  843  859  962  967  1010  1014  1031  1080
#>  [2,]    7    1  623  668  765  904 1057 1070 1072  1096     0     0     0
#>  [3,]    6    1  581  671  687  860  978 1019 1096     0     0     0     0
#>  [4,]   10    1  609  619  831  849  886  934  940   969  1024  1086  1096
#>  [5,]   13    1  566  628  698  777  784  822  876   879   890   982  1056
#>  [6,]    8    1  557  705  790  853  886  950 1067  1068  1096     0     0
#>  [7,]   10    1  671  856  858  907  947  968  969   997  1010  1078  1096
#>  [8,]    8    1  557  722  729  857  945  992 1025  1086  1096     0     0
#>  [9,]    9    1  645  834  837  875  896  981 1008  1023  1041  1096     0
#> [10,]    9    1  746  826  864  878  911  957  963  1032  1041  1096     0
#> [11,]   11    1  738  777  894  899  945  964 1008  1015  1044  1050  1083
#> [12,]    6    1  687  915  976  977 1000 1086 1096     0     0     0     0
#> [13,]    7    1  575  678  849  962 1008 1028 1061  1096     0     0     0
#> [14,]    7    1  794  851  923  981 1019 1034 1084  1096     0     0     0
#> [15,]   11    1  668  671  680  694  753  768  863   905   946   971   976
#> [16,]    9    1  678  841  944  996 1014 1031 1039  1049  1081  1096     0
#> [17,]   12    1  654  668  671  688  711  738  870   878   935  1001  1031
#> [18,]    9    1  571  687  697  704  803  966  982  1020  1086  1096     0
#> [19,]   10    1  598  615  659  785  856  885 1009  1026  1042  1079  1096
#> [20,]    9    1  415  784  791  839  872  879  887   897   985  1096     0
#> [21,]    7    1  666  794  828  935  969 1045 1086  1096     0     0     0
#> [22,]    9    1  830  842  892  924  934  971  983   997  1051  1096     0
#> [23,]    8    1  615  940  996 1015 1019 1025 1049  1066  1096     0     0
#> [24,]    8    1  840  876  883  944  950  987 1048  1066  1096     0     0
#> [25,]   10    1  488  561  663  777  834  846  968   977  1071  1074  1096
#> [26,]   10    1  562  698  832  979  982 1014 1018  1028  1060  1062  1096
#> [27,]    6    1  553  791  841  910  945 1046 1096     0     0     0     0
#> [28,]   10    1  590  609  656  697  698  716  790   853  1061  1071  1096
#> [29,]   10    1  803  877  884  910  928  936  969   999  1014  1027  1096
#> [30,]   12    1  553  633  791  813  830  901  933   962   983  1000  1015
#> [31,]    7    1  590  613  654  859  999 1002 1086  1096     0     0     0
#> [32,]    9    1  619  666  711  849  953  996 1007  1012  1030  1096     0
#> [33,]    8    1  618  846  945  954  970  999 1013  1028  1096     0     0
#> [34,]    9    1  614  697  771  790  875  886  900   972  1003  1096     0
#> [35,]    8    1  621  706  777  785  872  962 1039  1067  1096     0     0
#> [36,]   10    1  621  722  733  794  867  886  945  1046  1078  1079  1096
#> [37,]    8    1  646  649  803  870  877  971 1012  1061  1096     0     0
#> [38,]   11    1  553  621  698  774  993 1000 1012  1028  1041  1067  1086
#> [39,]    8    1  672  727  881  893  948  990  995  1083  1096     0     0
#> [40,]    9    1  415  583  725  894  948  961  984   995   997  1096     0
#> [41,]    9    1  647  728  768  826  837  860  901   946  1000  1096     0
#> [42,]    8    1  556  733  861  901  933  993 1000  1068  1096     0     0
#> [43,]   10    1  671  771  798  882  905  967  987   992  1038  1048  1096
#> [44,]    7    1  415  697  774  824  993 1038 1059  1096     0     0     0
#> [45,]   10    1  623  647  784  934  978 1013 1019  1024  1061  1070  1096
#> [46,]   10    1  618  830  849  887  969  999 1029  1056  1062  1083  1096
#> [47,]   12    1  654  704  830  877  921  954  971   994  1021  1029  1031
#> [48,]    8    1  556  605  856  886  946  948  989   991  1096     0     0
#> [49,]    7    1  621  861  901  970 1047 1055 1064  1096     0     0     0
#> [50,]    7    1  571  765  828  848  866 1025 1086  1096     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]  1083  1096     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]  1069  1071  1096     0     0     0     0
#>  [6,]     0     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]  1096     0     0     0     0     0     0
#> [12,]     0     0     0     0     0     0     0
#> [13,]     0     0     0     0     0     0     0
#> [14,]     0     0     0     0     0     0     0
#> [15,]  1096     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]  1076  1096     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]     0     0     0     0     0     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]     0     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]  1067  1096     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]     0     0     0     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]  1096     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0
#> [44,]     0     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]  1056  1096     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1   60  415  613  621  671  867  951  961  980  985 1001 1013 1024
#> [16] 1052 1054 1075 1076 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   16   56   59   97  126  132  147  582  660  808  838  861  866
#> [16]  937  962 1031 1048 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  379  388  423  477  575  615  691  739  899  911  924  936  937
#> [16]  969  977 1033 1068 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1  607  833  839  857  862  864  894  896  948  950  972  973 1004
#> [16] 1016 1018 1021 1064 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1  226  618  880  882  886  890  905   931   951   963   976
#>  [2,]   17    1  618  677  678  687  784  823  830   836   840   846   859
#>  [3,]   12    1  634  671  705  784  823  864  886   926   959   975   985
#>  [4,]   17    1  566  687  716  790  828  831  859   875   887   899   977
#>  [5,]   12    1  777  785  813  822  837  860  861   954   977   985  1035
#>  [6,]   17    1  628  634  705  729  809  946  951   970   974   986  1010
#>  [7,]   15    1  604  654  659  687  705  728  749   765   850   870   911
#>  [8,]   17    1  221  634  642  728  755  817  853   877   878   887   937
#>  [9,]   17    1  488  614  785  839  851  864  880   914   945  1001  1052
#> [10,]   17    1  654  657  741  810  838  895  898   960   972   988   994
#> [11,]   16    1   60  642  678  728  833  863  899   901   920   959  1005
#> [12,]   17    1  668  687  785  834  853  858  861   904   933   974   986
#> [13,]   17    1  488  562  619  634  753  809  823   910   941   951   969
#> [14,]   17    1  613  633  657  716  791  838  845   854   879   891   947
#> [15,]   17    1  649  713  716  790  851  876  911   977   989   995  1024
#> [16,]   17    1  221  488  609  716  727  813  844   856   943   961   972
#> [17,]   17    1  271  571  656  680  733  785  813   887   894   910   933
#> [18,]   17    1  656  657  725  780  826  848  881   893   945   966   968
#> [19,]   17    1  581  614  655  666  728  810  832   891   897   986   995
#> [20,]   17    1  663  694  774  841  848  892  899   910   928   931   934
#> [21,]   16    1  619  677  738  854  856  864  894   899   918   946   967
#> [22,]   17    1  561  571  619  677  688  771  850   852   899   903   938
#> [23,]   17    1  553  615  629  810  840  845  879   883   885   957   978
#> [24,]   17    1  415  621  717  738  777  798  888   895   899   933   951
#> [25,]   17    1  488  633  666  698  708  729  866   892   941   990   993
#> [26,]   17    1  697  698  728  761  765  771  823   832   846   865   904
#> [27,]   17    1   51  634  822  837  839  854  861   879   910   935   957
#> [28,]   14    1  556  688  780  810  822  839  850   915   982  1017  1037
#> [29,]   17    1  598  604  738  761  842  889  941   945   973   988   991
#> [30,]   16    1  666  680  722  763  771  809  846   936   937   953   959
#> [31,]   17    1  415  666  671  672  678  794  826   844   878   882   940
#> [32,]   17    1  571  681  697  728  826  831  846   849   878   898   935
#> [33,]   16    1   60  649  656  763  843  858  881   901   915   916   990
#> [34,]   17    1  354  562  619  647  749  755  763   845   864   901   907
#> [35,]   17    1  233  613  647  659  705  790  792   880   883   897   948
#> [36,]   17    1  618  628  629  649  668  727  791   861   892   989   995
#> [37,]   17    1   51  596  663  708  725  755  817   830   864   905   932
#> [38,]   17    1  271  562  571  798  823  834  853   870   905   915   924
#> [39,]   17    1  615  642  645  763  774  837  852   865   895   991  1004
#> [40,]   17    1  623  688  722  780  794  823  833   841   848   859   872
#> [41,]   14    1  621  647  666  727  824  828  878   881   904   934   936
#> [42,]   17    1   60  556  628  672  687  722  791   823   875   877   923
#> [43,]   17    1  354  415  598  619  666  719  872   877   895   963  1008
#> [44,]   17    1  571  621  623  647  834  872  884   937   966   971   990
#> [45,]   16    1  354  566  614  629  774  929  936   963   975  1006  1007
#> [46,]   12    1  666  719  755  780  809  813  822   836  1001  1022  1069
#> [47,]   17    1  671  694  741  809  828  851  856   860   861   937   982
#> [48,]   17    1  654  680  711  763  765  774  794   841   919   932   933
#> [49,]   15    1  649  708  780  798  846  864  904   944   969   992  1007
#> [50,]   17    1  556  561  598  763  771  790  862   884   894   897   910
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   990  1007  1022  1023  1039  1093  1096
#>  [2,]   904   915   919   982  1085  1095  1096
#>  [3,]  1003  1096     0     0     0     0     0
#>  [4,]  1013  1019  1030  1044  1051  1091  1096
#>  [5,]  1049  1096     0     0     0     0     0
#>  [6,]  1023  1047  1072  1081  1084  1087  1096
#>  [7,]   946   954  1010  1011  1096     0     0
#>  [8,]   941   948   961  1057  1066  1070  1096
#>  [9,]  1059  1065  1066  1087  1088  1092  1096
#> [10,]  1008  1017  1030  1054  1072  1088  1096
#> [11,]  1045  1049  1065  1069  1088  1096     0
#> [12,]  1035  1051  1057  1058  1059  1093  1096
#> [13,]   974  1020  1032  1058  1061  1084  1096
#> [14,]   964   987   994  1052  1063  1091  1096
#> [15,]  1030  1032  1034  1073  1078  1087  1096
#> [16,]   977   980   982  1003  1017  1046  1096
#> [17,]   943   989  1007  1051  1082  1095  1096
#> [18,]   969  1050  1059  1075  1076  1088  1096
#> [19,]  1003  1004  1022  1046  1055  1061  1096
#> [20,]   940   991  1000  1054  1079  1085  1096
#> [21,]  1002  1017  1018  1028  1073  1096     0
#> [22,]   974  1013  1014  1037  1047  1071  1096
#> [23,]   991  1000  1023  1053  1076  1080  1096
#> [24,]   973   981   988  1040  1055  1083  1096
#> [25,]  1008  1015  1021  1033  1045  1061  1096
#> [26,]   960   986  1003  1035  1039  1080  1096
#> [27,]   972   976   989  1008  1033  1039  1096
#> [28,]  1046  1062  1064  1096     0     0     0
#> [29,]   999  1009  1013  1022  1033  1081  1096
#> [30,]   981   990  1050  1063  1069  1096     0
#> [31,]   944   963   969   972   976  1074  1096
#> [32,]   936   937   950   970  1016  1037  1096
#> [33,]  1002  1019  1061  1079  1090  1096     0
#> [34,]   948   997  1001  1048  1053  1068  1096
#> [35,]   968   985   996  1029  1056  1066  1096
#> [36,]  1012  1013  1034  1044  1045  1055  1096
#> [37,]   949   961   967  1054  1065  1068  1096
#> [38,]   941   979   997  1011  1030  1039  1096
#> [39,]  1020  1025  1049  1051  1071  1087  1096
#> [40,]   912   939   944   953  1017  1035  1096
#> [41,]   999  1052  1062  1096     0     0     0
#> [42,]   937   947   991  1013  1064  1074  1096
#> [43,]  1009  1017  1033  1049  1086  1092  1096
#> [44,]  1019  1026  1064  1069  1073  1085  1096
#> [45,]  1018  1029  1036  1071  1072  1096     0
#> [46,]  1071  1096     0     0     0     0     0
#> [47,]   999  1006  1023  1033  1052  1088  1096
#> [48,]   945   971   988   998  1022  1090  1096
#> [49,]  1010  1025  1048  1084  1096     0     0
#> [50,]   932   936   998  1073  1084  1085  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986