Generate a list of candidate changepoints using a genetic algorithm

Generate a list of candidate changepoints using a genetic algorithm

Usage

evolve_gbmdl(x, mat_cp, these_bmdls)

junta_1_puntos_cambio(padres, mat_cp)

junta_k_puntos_cambio(mat_padres, mat_cp)

mata_1_tau_volado(cp, prob_volado = 0.5)

mata_k_tau_volado(mat_cp)

muta_1_cp_BMDL(
  cp,
  x,
  probs_nuevos_muta0N = c(0.8, 0.1, 0.1),
  dist_extremos = 10,
  min_num_cpts = 1,
  mutation_possibilities = c(-1, 0, 1),
  mutation_probs = c(0.3, 0.4, 0.3),
  max_num_cp = 20
)

muta_k_cp_BMDL(mat_cp, x)

sim_1_cp_BMDL(x, max_num_cp = 20, prob_inicial = 0.06)

sim_k_cp_BMDL(x, generation_size = 50, max_num_cp = 20)

probs_vec_MDL(vec_MDL, probs_rank0_MDL1 = 0)

selec_k_pares_de_padres(vec_probs)

chromo2tau(chromo)

mat_cp_2_list(mat_cp)

Arguments

x

A time series object

mat_cp

A matrix of potential changepoints

these_bmdls

vector of BMDL() scores

padres

vector de longitud dos con índice de papa e índice de mama

mat_padres

matriz de kx2 la cual contiene en sus renglones las parejas de padres

cp

vector cromosoma que se va a poner a prueba

prob_volado

probabilidad de quitar un tiempo de cambio existente utilizado por mata_k_tau_volado para quitar elementos de más. Se recomienda dejar el valor de 0.5 ya que así al juntar los pc del padre y madre se eliminará la mitad de estos

probs_nuevos_muta0N

probabilidades de mutar 0,1,2,...,l hasta cierto numero l; eg si vale c(.5,.2,.2,.1) se tiene una probabilidad 0.5 de mutar 0 (de no mutar), probabilidad 0.2 de mutar 1,, probabilidad 0.2 de mutar 2, y, probabilidad 0.1 de mutar 3.

dist_extremos

distancia entre el primer los puntos de cambio v_0 y v_1 al igual que entre v_m y v_{m+1}; distancia minima que debe de haber de un punto de cambio y los valores 1 y T, donde T es la longitud total de la serie

min_num_cpts

es la cota inferior del número de puntos de cambio que puede tener un cromosoma

mutation_possibilities

vector con mutaciones posibles; eg si mutaciones=c(-1,0,1) entonces un punto de cambio puede ser movido una unidad a la izquierda, puede quedarse igual, o moverse una unidad a la derecha

mutation_probs

probabilidades de mutación. Las longitudes de este vector y mutaciones tienen que ser iguales; eg si mutaciones=c(-1,0,1) y probs_muta = c(.2, .6, .2) entonces se tiene una probabilidad .2 de que el punto de cambio se desplace a la izquierda, probabilidad .6 de quedar igual, y probabilidad . 2 de ser movido a la derecha

max_num_cp

el máximo número de rebases. Este parámetro se ocupa en particular para que todos los cromosomas quepan en una matriz.

prob_inicial

probabilidad de que en la primera generación un punto cualquiera sea punto de cambio. Se recomienda =.5 ya que con esto se distribuyen relativamente uniformes los puntos de cambio

generation_size

tamaño de las generaciones

vec_MDL

vector con valores MDL

OBSERVACIÓN: Esto regresa numeros negativos, los cuales mientras más negativo mejor, ya que dará que es un mejor vector de tiempos de cambio. Es decir, un MDL de -6000 es mejor que -4000

probs_rank0_MDL1

para medir obtener la probabilidad de los padres se pueden tomar o las probabilidades con respecto a los rangos (como en el artículo) o se pueden tomar las probabilidades con respecto a el MDL. La diferencia radica en que si se toma con respecto al MDL se tendrá que un cromosoma con un gran MDL este tendrá una gran ventaja de ocurrir, en cambio cuando solo se tiene rank esta ventaja gran ventaja se reduce

vec_probs

vector de probabilidades de selección de cada uno de los cromosomas

chromo

Chromosome, from a row of the matrix mat_cp

Value

regresa una matriz de las mismas dimensiones que mat_cp, pero con los nuevos cromosomas

el mismo cromosoma sin algunos de sus puntos de cambio

regresa una matriz a la cual se le quitaron a sus cromosomas algunos puntos de cambio

regresa un vector mutado

regreas una mat_cp mutada

regresa una matriz de k por max_num_cp+3, la cual en cada renglón tiene una simulación de un vector de tiempos de cambio

regresa un vector de probabilidades

Details

regresa un vector de tamaño max_num_cp+3 donde la primera entrada es m, la segunda \(v_0=1, ...., v_{m+1}=N,0,...,0\)

por ejemplo: c(4,1,3,8,11,15,20,0,0,0,0) para m=4, \(max\_num\_cp=8\), \(N=20\). Se tienen m puntos de cambio, los cuales \(\tau_0=1\) y \(\tau_{m+1}= N+1\), pero en nuestro caso tenemos que los vectores cp tienen \(c(m,\tau_0=1,\tau_1,...,\tau_{m-1},\tau_m= N,0,0,0)\) por lo cual se nosotros:

empieza con el número de puntos de cambio;

la segunda entrada es un uno;

la tercera entrada es el primer punto de cambio;

las siguientes son otros puntos de cambio;

la siguiente entrada después de punto de cambio tiene el valor N; y

los siguientes son númores cero hasta llenarlo para que sea de tamaño max_num_cp

Examples

mat_cp <- sim_k_cp_BMDL(DataCPSim)
bmdls <- mat_cp |> 
  mat_cp_2_list() |> 
  evaluate_cpts(.data = as.ts(DataCPSim), model_fn = fit_nhpp) |> 
  dplyr::pull(BMDL)
evolve_gbmdl(exceedances(DataCPSim), mat_cp, bmdls)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   10    1  575  623  633  713  741  777  828   920   936  1019  1096
#>  [2,]   10    1  649  728  825  849 1001 1012 1026  1042  1047  1055  1096
#>  [3,]    7    1  713  761  832  838 1063 1074 1086  1096     0     0     0
#>  [4,]    8    1  571  628  741  749  771  903  957   990  1096     0     0
#>  [5,]    9    1  415  642  728  860  879 1023 1030  1036  1086  1096     0
#>  [6,]   11    1  415  825  853  856  860  891  969  1004  1018  1038  1069
#>  [7,]    5    1  596  753  837  951  969 1096    0     0     0     0     0
#>  [8,]    8    1  791  794  798  851  955  969 1038  1052  1096     0     0
#>  [9,]    8    1  415  556  688 1005 1042 1043 1047  1053  1096     0     0
#> [10,]   10    1  488  571  657  753  881  894  929   936   953  1086  1096
#> [11,]   11    1  415  711  822  974  975 1011 1012  1032  1034  1045  1077
#> [12,]   11    1  415  729  746  859  963 1001 1026  1043  1065  1071  1086
#> [13,]   11    1  561  716  741  794  838  841  853   892   955  1007  1038
#> [14,]   12    1  609  659  668  768  851  884  887   897   919  1000  1031
#> [15,]   10    1  634  746  825  826  862  933  945   966   974  1065  1096
#> [16,]    6    1  556  722  741  860  904 1062 1096     0     0     0     0
#> [17,]    9    1  761  774  822  915  916  926  936   974  1079  1096     0
#> [18,]    7    1  415  618  790  851  865  988 1068  1096     0     0     0
#> [19,]    9    1  589  621  704  861  997 1001 1039  1062  1086  1096     0
#> [20,]   12    1  415  561  598  645  666  746  765   838   858   872   988
#> [21,]    8    1  589  633  798  856  941  970 1010  1078  1096     0     0
#> [22,]    9    1  575  774  870  928  930  948  970  1004  1062  1096     0
#> [23,]    6    1  590  615  621  886  979 1008 1096     0     0     0     0
#> [24,]    9    1  548  621  678  733  813  836  993  1044  1061  1096     0
#> [25,]    8    1  614  678  837  926  946  980 1005  1007  1096     0     0
#> [26,]   10    1  561  609  774  826  877  894  912   979   980   992  1096
#> [27,]   10    1  697  839  918  949  957  994 1004  1015  1070  1076  1096
#> [28,]   10    1  642  831  859  900  946  953  991   996  1025  1086  1096
#> [29,]    7    1  830  859  894  916  992 1044 1086  1096     0     0     0
#> [30,]    9    1  609  706  733  934  985  990 1033  1051  1086  1096     0
#> [31,]    8    1  561  794  871  932  935 1002 1080  1086  1096     0     0
#> [32,]    9    1  623  633  774  880  887  897  931   935  1031  1096     0
#> [33,]    7    1  415  725  738  803  817  840 1086  1096     0     0     0
#> [34,]   13    1  415  561  605  621  633  649  777   856   944   979   994
#> [35,]    6    1  656  719  878  974 1008 1057 1096     0     0     0     0
#> [36,]    9    1  605  694  790  798  854  980 1016  1060  1067  1096     0
#> [37,]    7    1  590  619  706  717  780  810 1045  1096     0     0     0
#> [38,]   11    1  488  628  694  719  876  887  963   985  1019  1045  1058
#> [39,]   10    1  609  763  838  845  923  928  947   961   975  1064  1096
#> [40,]    9    1  566  583  621  725  901 1041 1061  1078  1086  1096     0
#> [41,]    9    1  697  705  780  798  859  883 1068  1078  1086  1096     0
#> [42,]    6    1  589  634  666  932  996 1084 1096     0     0     0     0
#> [43,]    8    1  634  677  826  856  921  969 1051  1067  1096     0     0
#> [44,]   11    1  589  628  649  677  892  900  919   928  1024  1046  1086
#> [45,]    7    1  415  598  645  668  727  852  963  1096     0     0     0
#> [46,]    8    1  575  618  733  919  990 1023 1061  1086  1096     0     0
#> [47,]   10    1  415  556  561  621  722  883  898  1031  1064  1086  1096
#> [48,]   10    1  645  798  826  871  885  920  929   954  1011  1068  1096
#> [49,]    8    1  415  645  657  785  879  914  930  1045  1096     0     0
#> [50,]    7    1  596  663  857  885 1049 1067 1070  1096     0     0     0
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]     0     0     0     0     0     0     0
#>  [2,]     0     0     0     0     0     0     0
#>  [3,]     0     0     0     0     0     0     0
#>  [4,]     0     0     0     0     0     0     0
#>  [5,]     0     0     0     0     0     0     0
#>  [6,]  1096     0     0     0     0     0     0
#>  [7,]     0     0     0     0     0     0     0
#>  [8,]     0     0     0     0     0     0     0
#>  [9,]     0     0     0     0     0     0     0
#> [10,]     0     0     0     0     0     0     0
#> [11,]  1096     0     0     0     0     0     0
#> [12,]  1096     0     0     0     0     0     0
#> [13,]  1096     0     0     0     0     0     0
#> [14,]  1069  1096     0     0     0     0     0
#> [15,]     0     0     0     0     0     0     0
#> [16,]     0     0     0     0     0     0     0
#> [17,]     0     0     0     0     0     0     0
#> [18,]     0     0     0     0     0     0     0
#> [19,]     0     0     0     0     0     0     0
#> [20,]  1005  1096     0     0     0     0     0
#> [21,]     0     0     0     0     0     0     0
#> [22,]     0     0     0     0     0     0     0
#> [23,]     0     0     0     0     0     0     0
#> [24,]     0     0     0     0     0     0     0
#> [25,]     0     0     0     0     0     0     0
#> [26,]     0     0     0     0     0     0     0
#> [27,]     0     0     0     0     0     0     0
#> [28,]     0     0     0     0     0     0     0
#> [29,]     0     0     0     0     0     0     0
#> [30,]     0     0     0     0     0     0     0
#> [31,]     0     0     0     0     0     0     0
#> [32,]     0     0     0     0     0     0     0
#> [33,]     0     0     0     0     0     0     0
#> [34,]  1046  1086  1096     0     0     0     0
#> [35,]     0     0     0     0     0     0     0
#> [36,]     0     0     0     0     0     0     0
#> [37,]     0     0     0     0     0     0     0
#> [38,]  1096     0     0     0     0     0     0
#> [39,]     0     0     0     0     0     0     0
#> [40,]     0     0     0     0     0     0     0
#> [41,]     0     0     0     0     0     0     0
#> [42,]     0     0     0     0     0     0     0
#> [43,]     0     0     0     0     0     0     0
#> [44,]  1096     0     0     0     0     0     0
#> [45,]     0     0     0     0     0     0     0
#> [46,]     0     0     0     0     0     0     0
#> [47,]     0     0     0     0     0     0     0
#> [48,]     0     0     0     0     0     0     0
#> [49,]     0     0     0     0     0     0     0
#> [50,]     0     0     0     0     0     0     0
sim_1_cp_BMDL(exceedances(DataCPSim))
#>  [1]   17    1   51  645  746  771  798  828  831  833  877  885  940 1015 1042
#> [16] 1043 1056 1065 1092 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_1))
#>  [1]   17    1   12   34  167  258  275  294  358  800  820  835  900  925  942
#> [16]  967  994  997 1075 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_2))
#>  [1]   17    1  444  489  607  620  748  811  824  833  849  863  894  913  991
#> [16] 1018 1027 1073 1085 1096
sim_1_cp_BMDL(exceedances(rlnorm_ts_3))
#>  [1]   17    1  586  590  592  649  697  766  767  808  864  880  881  943  949
#> [16] 1035 1049 1058 1088 1096
sim_1_cp_BMDL(exceedances(bogota_pm))
#>  [1]   17    1   40   46   47   59   87  186  395  424  478  558  617  712  792
#> [16]  804  934 1033 1093 1096

sim_k_cp_BMDL(DataCPSim)
#>       [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>  [1,]   17    1   20  562  613  633  687  717  746   826   851   881   908
#>  [2,]   14    1   51  798  875  898  905  908  936   953   977   989  1038
#>  [3,]   17    1  633  830  842  859  863  866  898   905   914   929   951
#>  [4,]   17    1  488  566  719  810  832  863  864   905   918   924   995
#>  [5,]   17    1  354  633  708  727  763  777  833   900   915   946  1015
#>  [6,]   17    1  711  738  741  763  790  848  883   886   889   907   914
#>  [7,]   17    1  488  581  654  666  738  856  858   862   865   886   891
#>  [8,]   17    1  590  706  831  844  851  852  881   934   938   943   976
#>  [9,]   17    1  548  581  604  657  678  688  765   817   823   893   948
#> [10,]   17    1  271  308  553  761  765  828  839   845   888   899   924
#> [11,]   17    1  271  308  629  654  704  755  765   933   934   954   969
#> [12,]   17    1  618  688  716  780  825  849  864   870   910   924   935
#> [13,]   17    1  614  649  823  824  868  876  903   914   926   933   949
#> [14,]   17    1  647  657  733  749  809  914  921   924   929   945   969
#> [15,]   17    1  614  708  716  755  785  822  862   903   916   923   954
#> [16,]   16    1   60  271  566  633  678  809  842   864   881   890   914
#> [17,]   17    1  571  629  654  687  727  790  850   856   870   884   915
#> [18,]   17    1  613  688  717  725  753  784  833   853   938   982   983
#> [19,]   17    1  226  233  561  654  719  727  763   868   875   924   971
#> [20,]   14    1   20  633  727  824  856  865  871   893   895   987   996
#> [21,]   17    1  233  308  575  655  698  748  791   848   872   912   938
#> [22,]   16    1  415  618  621  672  725  809  822   851   853   862   879
#> [23,]   15    1  596  642  649  677  824  853  932   945   975   987   995
#> [24,]   17    1  596  609  668  704  774  846  848   886   934   970   995
#> [25,]   17    1  583  598  621  628  687  694  768   866   883   911   966
#> [26,]   17    1  596  628  647  798  810  825  851   876   919   920   921
#> [27,]   17    1   60  680  694  768  825  840  882   892   897   907   911
#> [28,]   17    1   60  556  748  866  867  875  879   888   893   901   946
#> [29,]   17    1  354  566  621  755  803  832  867   889   918   945   961
#> [30,]   17    1  748  771  867  875  882  920  926   955   960   968   981
#> [31,]   17    1  581  649  656  763  871  943  957   959  1005  1009  1032
#> [32,]   17    1  609  618  655  706  716  771  791   845   860   871   879
#> [33,]   13    1  663  694  717  931  971  999 1002  1018  1029  1030  1034
#> [34,]   17    1  354  596  649  729  777  784  860   878   880   905   947
#> [35,]   17    1  221  688  716  828  843  856  868   871   928   941   971
#> [36,]   17    1  629  642  656  777  833  949  954   967   982   993  1009
#> [37,]   17    1  271  657  668  746  840  857  891   934   945   946   985
#> [38,]   17    1  645  663  727  867  872  920  935   937   940   943   948
#> [39,]   15    1  556  833  846  883  889  934  954   972   978   981   998
#> [40,]   17    1  226  562  738  833  841  857  859   868   880   887   890
#> [41,]   17    1  649  768  813  878  888  891  895   904   945   950   989
#> [42,]   14    1  708  825  826  833  844  853  863   884   910   993  1025
#> [43,]   16    1  226  548  628  671  681  722  872   930   968   969   972
#> [44,]   17    1  645  657  728  761  825  850  862   879   898   967  1001
#> [45,]   17    1  233  761  837  839  851  858  879   889   893   910   929
#> [46,]   17    1  605  655  656  836  841  878  882   901   908   939   941
#> [47,]   17    1  221  590  605  722  725  753  838   881   888   912   924
#> [48,]   17    1  308  575  618  722  738  794  872   882   907   946   950
#> [49,]   16    1  548  581  656  708  722  822  889   912   944   987   991
#> [50,]   17    1  654  657  704  738  790  825  844   919   967   971  1001
#>       [,14] [,15] [,16] [,17] [,18] [,19] [,20]
#>  [1,]   939   945   947   991  1052  1082  1096
#>  [2,]  1044  1050  1066  1096     0     0     0
#>  [3,]   961   967  1038  1071  1076  1092  1096
#>  [4,]  1011  1019  1033  1060  1074  1094  1096
#>  [5,]  1019  1020  1026  1045  1052  1063  1096
#>  [6,]   941  1005  1038  1046  1076  1092  1096
#>  [7,]   899   973   987  1024  1039  1084  1096
#>  [8,]   979   997  1002  1054  1065  1077  1096
#>  [9,]   962   967   991  1023  1058  1088  1096
#> [10,]   957   968   999  1015  1033  1065  1096
#> [11,]   970   973  1021  1034  1083  1093  1096
#> [12,]   964  1005  1057  1058  1089  1094  1096
#> [13,]   964   967   994  1015  1075  1081  1096
#> [14,]   984   989  1008  1021  1062  1085  1096
#> [15,]  1009  1018  1038  1058  1075  1081  1096
#> [16,]   920   971  1002  1031  1082  1096     0
#> [17,]   933   996  1006  1050  1059  1080  1096
#> [18,]   984  1013  1017  1028  1032  1034  1096
#> [19,]  1028  1039  1049  1066  1069  1088  1096
#> [20,]  1024  1032  1091  1096     0     0     0
#> [21,]   947   953   963   972  1033  1092  1096
#> [22,]   945  1008  1083  1087  1090  1096     0
#> [23,]  1004  1018  1042  1075  1096     0     0
#> [24,]  1009  1012  1022  1029  1046  1056  1096
#> [25,]   967   970  1001  1015  1061  1094  1096
#> [26,]   929   948   949   950  1067  1088  1096
#> [27,]   921   977  1006  1049  1075  1092  1096
#> [28,]   987   996  1001  1010  1022  1053  1096
#> [29,]   983   990  1023  1027  1042  1080  1096
#> [30,]   984   997  1040  1049  1077  1086  1096
#> [31,]  1035  1046  1056  1064  1089  1090  1096
#> [32,]   936   986   991   998  1040  1056  1096
#> [33,]  1046  1057  1096     0     0     0     0
#> [34,]   988   993  1046  1049  1055  1078  1096
#> [35,]   980  1002  1015  1044  1070  1086  1096
#> [36,]  1011  1018  1059  1064  1065  1088  1096
#> [37,]  1006  1021  1065  1068  1071  1077  1096
#> [38,]  1004  1019  1023  1041  1074  1085  1096
#> [39,]  1033  1035  1051  1065  1096     0     0
#> [40,]   946   959   977  1069  1073  1091  1096
#> [41,]  1034  1038  1052  1061  1078  1095  1096
#> [42,]  1032  1063  1088  1096     0     0     0
#> [43,]  1042  1063  1064  1072  1095  1096     0
#> [44,]  1009  1032  1034  1043  1045  1079  1096
#> [45,]   953   957   960  1005  1010  1083  1096
#> [46,]   991  1003  1018  1024  1063  1080  1096
#> [47,]   987   992  1015  1040  1075  1090  1096
#> [48,]  1008  1038  1043  1058  1080  1082  1096
#> [49,]  1008  1013  1016  1037  1051  1096     0
#> [50,]  1037  1058  1061  1075  1082  1090  1096

chromo <- c(4, 1, 557, 877 , 905, 986, 1096, 0, 0, 0)
chromo2tau(chromo)
#> [1] 557 877 905 986